miri
Module alloc_addresses

This module is responsible for managing the absolute addresses that allocations are located at, and for casting between pointers and integers based on those addresses.

Modules

reuse_pool 🔒

Manages a pool of addresses that can be reused.

Structs

GlobalStateInner

Enums

ProvenanceMode

Traits

EvalContextExt

EvalContextExtPriv 🔒

Functions

align_addr 🔒

Shifts addr to make it aligned with align by rounding addr to the smallest multiple of align that is larger or equal to addr

Type Aliases

GlobalState