miri::alloc_addresses::reuse_pool
Struct ReusePool

pub struct ReusePool {
    address_reuse_rate: f64,
    address_reuse_cross_thread_rate: f64,
    pool: Vec<Vec<(u64, Size, ThreadId, VClock)>>,
}

The pool strikes a balance between exploring more possible executions and making it more likely to find bugs. The hypothesis is that bugs are more likely to occur when reuse happens for allocations with the same layout, since that can trigger e.g. ABA issues in a concurrent data structure. Therefore we only reuse allocations when size and alignment match exactly.

Fields
address_reuse_rate: f64address_reuse_cross_thread_rate: f64pool: Vec<Vec<(u64, Size, ThreadId, VClock)>>

The i-th element in pool stores allocations of alignment 2^i. We store these reusable allocations as address-size pairs, the list must be sorted by the size and then the thread ID.

Each of these maps has at most MAX_POOL_SIZE elements, and since alignment is limited to less than 64 different possible value, that bounds the overall size of the pool.

We also store the ID and the data-race clock of the thread that donated this pool element, to ensure synchronization with the thread that picks up this address.

Implementations

impl ReusePool

pub fn new(config: &MiriConfig) -> Self

fn subpool(&mut self, align: Align) -> &mut Vec<(u64, Size, ThreadId, VClock)>

pub fn add_addr( &mut self, rng: &mut impl Rng, addr: u64, size: Size, align: Align, kind: MemoryKind, thread: ThreadId, clock: impl FnOnce() -> VClock, )

pub fn take_addr( &mut self, rng: &mut impl Rng, size: Size, align: Align, kind: MemoryKind, thread: ThreadId, ) -> Option<(u64, Option<VClock>)>

Returns the address to use and optionally a clock we have to synchronize with.