miri::borrow_tracker::stacked_borrows::diagnostics
Struct DiagnosticCx

pub struct DiagnosticCx<'history, 'ecx, 'tcx> {
    operation: Operation,
    machine: &'ecx MiriMachine<'tcx>,
    history: &'history mut AllocHistory,
    offset: Size,
}

Fields
operation: Operationmachine: &'ecx MiriMachine<'tcx>history: &'history mut AllocHistoryoffset: Size
Implementations

impl<'history, 'ecx, 'tcx> DiagnosticCx<'history, 'ecx, 'tcx>

pub fn unbuild(self) -> DiagnosticCxBuilder<'ecx, 'tcx>

impl<'history, 'ecx, 'tcx> DiagnosticCx<'history, 'ecx, 'tcx>

pub fn start_grant(&mut self, perm: Permission)

pub fn log_creation(&mut self)

pub fn log_invalidation(&mut self, tag: BorTag)

pub fn log_protector(&mut self)

pub fn get_logs_relevant_to( &self, tag: BorTag, protector_tag: Option<BorTag>, ) -> Option<TagHistory>

pub(super) fn grant_error(&self, stack: &Stack) -> InterpErrorKind<'tcx>

Report a descriptive error when new could not be granted from derived_from.

pub(super) fn access_error(&self, stack: &Stack) -> InterpErrorKind<'tcx>

Report a descriptive error when access is not permitted based on tag.

pub(super) fn protector_error( &self, item: &Item, kind: ProtectorKind, ) -> InterpErrorKind<'tcx>

pub fn dealloc_error(&self, stack: &Stack) -> InterpErrorKind<'tcx>

pub fn check_tracked_tag_popped(&self, item: &Item, global: &GlobalStateInner)