Trait miri::borrow_tracker::stacked_borrows::EvalContextPrivExt

trait EvalContextPrivExt<'mir: 'ecx, 'tcx: 'mir, 'ecx>: MiriInterpCxExt<'mir, 'tcx> {
    fn sb_reborrow(
        &mut self,
        place: &MPlaceTy<'tcx, Provenance>,
        size: Size,
        new_perm: NewPermission,
        new_tag: BorTag,
        retag_cause: RetagCause
    ) -> InterpResult<'tcx, Option<AllocId>> { ... }
    fn sb_retag_reference(
        &mut self,
        val: &ImmTy<'tcx, Provenance>,
        new_perm: NewPermission,
        cause: RetagCause
    ) -> InterpResult<'tcx, ImmTy<'tcx, Provenance>> { ... }
}

Provided Methods

fn sb_reborrow(
&mut self,
place: &MPlaceTy<'tcx, Provenance>,
size: Size,
new_perm: NewPermission,
new_tag: BorTag,
retag_cause: RetagCause
) -> InterpResult<'tcx, Option<AllocId>>

Returns the AllocId the reborrow was done in, if some actual borrow stack manipulation happened.

fn sb_retag_reference(
&mut self,
val: &ImmTy<'tcx, Provenance>,
new_perm: NewPermission,
cause: RetagCause
) -> InterpResult<'tcx, ImmTy<'tcx, Provenance>>

Retags an indidual pointer, returning the retagged version. kind indicates what kind of reference is being created.