Type Alias miri::borrow_tracker::stacked_borrows::AllocState

pub type AllocState = Stacks;

Aliased Type

struct AllocState {
    stacks: RangeMap<Stack>,
    history: AllocHistory,
    exposed_tags: HashSet<BorTag, BuildHasherDefault<FxHasher>>,
    modified_since_last_gc: bool,
}

Fields
stacks: RangeMap<Stack>history: AllocHistory

Stores past operations on this allocation

exposed_tags: HashSet<BorTag, BuildHasherDefault<FxHasher>>

The set of tags that have been exposed inside this allocation.

modified_since_last_gc: bool

Whether this memory has been modified since the last time the tag GC ran

Implementations

impl Stacks

Integration with the BorTag garbage collector

pub fn remove_unreachable_tags(&mut self, live_tags: &FxHashSet<BorTag>)

impl<'tcx> Stacks

Map per-stack operations to higher-level per-location-range operations.

fn new( size: Size, perm: Permission, tag: BorTag, id: AllocId, machine: &MiriMachine<'_, '_> ) -> Self

Creates a new stack with an initial tag. For diagnostic purposes, we also need to know the AllocId of the allocation this is associated with.

fn for_each( &mut self, range: AllocRange, dcx_builder: DiagnosticCxBuilder<'_, '_, 'tcx>, f: impl FnMut(&mut Stack, &mut DiagnosticCx<'_, '_, '_, 'tcx>, &mut FxHashSet<BorTag>) -> InterpResult<'tcx> ) -> InterpResult<'tcx>

Call f on every stack in the range.

impl Stacks

Glue code to connect with Miri Machine Hooks

pub fn new_allocation( id: AllocId, size: Size, state: &mut GlobalStateInner, kind: MemoryKind<MiriMemoryKind>, machine: &MiriMachine<'_, '_> ) -> Self

pub fn before_memory_read<'tcx, 'mir, 'ecx>( &mut self, alloc_id: AllocId, tag: ProvenanceExtra, range: AllocRange, machine: &'ecx MiriMachine<'mir, 'tcx> ) -> InterpResult<'tcx>where 'tcx: 'ecx,

pub fn before_memory_write<'tcx>( &mut self, alloc_id: AllocId, tag: ProvenanceExtra, range: AllocRange, machine: &MiriMachine<'_, 'tcx> ) -> InterpResult<'tcx>

pub fn before_memory_deallocation<'tcx>( &mut self, alloc_id: AllocId, tag: ProvenanceExtra, range: AllocRange, machine: &MiriMachine<'_, 'tcx> ) -> InterpResult<'tcx>