Struct miri::borrow_tracker::GlobalStateInner

pub struct GlobalStateInner {
    pub borrow_tracker_method: BorrowTrackerMethod,
    pub next_ptr_tag: BorTag,
    pub base_ptr_tags: FxHashMap<AllocId, BorTag>,
    pub next_call_id: CallId,
    pub protected_tags: FxHashMap<BorTag, ProtectorKind>,
    pub tracked_pointer_tags: FxHashSet<BorTag>,
    pub tracked_call_ids: FxHashSet<CallId>,
    pub retag_fields: RetagFields,
}

Extra global state, available to the memory access hooks.

Fields
borrow_tracker_method: BorrowTrackerMethod

Borrow tracker method currently in use.

next_ptr_tag: BorTag

Next unused pointer ID (tag).

base_ptr_tags: FxHashMap<AllocId, BorTag>

Table storing the “base” tag for each allocation. The base tag is the one used for the initial pointer. We need this in a separate table to handle cyclic statics.

next_call_id: CallId

Next unused call ID (for protectors).

protected_tags: FxHashMap<BorTag, ProtectorKind>

All currently protected tags. An item is protected if its tag is in this set, and it has the “protected” bit set. We add tags to this when they are created with a protector in reborrow, and we remove tags from this when the call which is protecting them returns, in GlobalStateInner::end_call. See Stack::item_popped for more details.

tracked_pointer_tags: FxHashSet<BorTag>

The pointer ids to trace

tracked_call_ids: FxHashSet<CallId>

The call ids to trace

retag_fields: RetagFields

Whether to recurse into datatypes when searching for pointers to retag.

Implementations

impl GlobalStateInner

Utilities for initialization and ID generation

pub fn new(
borrow_tracker_method: BorrowTrackerMethod,
tracked_pointer_tags: FxHashSet<BorTag>,
tracked_call_ids: FxHashSet<CallId>,
retag_fields: RetagFields
) -> Self

pub fn new_ptr(&mut self) -> BorTag

Generates a new pointer tag. Remember to also check track_pointer_tags and log its creation!

pub fn new_frame(&mut self, machine: &MiriMachine<'_, '_>) -> FrameState

pub fn end_call(&mut self, frame: &FrameExtra<'_>)

pub fn base_ptr_tag(
&mut self,
id: AllocId,
machine: &MiriMachine<'_, '_>
) -> BorTag

impl GlobalStateInner

pub fn new_allocation(
&mut self,
id: AllocId,
alloc_size: Size,
kind: MemoryKind<MiriMemoryKind>,
machine: &MiriMachine<'_, '_>
) -> AllocState