miri::borrow_tracker::tree_borrows::diagnostics
Struct DisplayFmt

struct DisplayFmt {
    wrapper: DisplayFmtWrapper,
    perm: DisplayFmtPermission,
    padding: DisplayFmtPadding,
    accessed: DisplayFmtAccess,
}

All parameters to determine how the tree is formatted.

Fields
wrapper: DisplayFmtWrapperperm: DisplayFmtPermissionpadding: DisplayFmtPaddingaccessed: DisplayFmtAccess
Implementations

impl DisplayFmt

fn print_perm(&self, perm: Option<LocationState>) -> String

Print the permission with the format Res/ Re*/ Act/ Frz/ Dis for accessed locations and ?Res/?Re*/?Act/?Frz/?Dis for unaccessed locations.

fn print_tag(&self, tag: BorTag, name: &Option<String>) -> String

Print the tag with the format <XYZ> if the tag is unnamed, and <XYZ=name> if the tag is named.

fn print_protector(&self, protector: Option<&ProtectorKind>) -> &'static str

Print extra text if the tag has a protector.