Struct miri::borrow_tracker::tree_borrows::diagnostics::DisplayFmtPadding

struct DisplayFmtPadding {
    join_middle: &'static str,
    join_last: &'static str,
    join_haschild: &'static str,
    join_default: &'static str,
    indent_middle: &'static str,
    indent_last: &'static str,
}

Formating of the tree structure.

Example:

DisplayFmtPadding {
    join_middle: "|-",
    join_last: "'-",
    join_haschild: "-+-",
    join_default: "---",
    indent_middle: "| ",
    indent_last: "  ",
}

will show the tree as

-+- root
 |--+- a
 |  '--+- b
 |     '---- c
 |--+- d
 |  '---- e
 '---- f

Fields
join_middle: &'static str

Connector for a child other than the last.

join_last: &'static str

Connector for the last child. Should have the same width as join_middle.

join_haschild: &'static str

Connector for a node that itself has a child.

join_default: &'static str

Connector for a node that does not have a child. Should have the same width as join_haschild.

indent_middle: &'static str

Indentation when there is a next child.

indent_last: &'static str

Indentation for the last child.