Struct miri::borrow_tracker::tree_borrows::perms::PermTransition

pub struct PermTransition {
    from: PermissionPriv,
    to: PermissionPriv,
}

Transition from one permission to the next.

Fields
from: PermissionPrivto: PermissionPriv
Implementations

impl PermTransition

pub fn summary(&self) -> &'static str

Readable explanation of the consequences of an event. Fits in the sentence “This accessed caused {trans.summary()}”.

Important: for the purposes of this explanation, Reserved is considered to have write permissions, because that’s what the diagnostics care about (otherwise Reserved -> Frozen would be considered a noop).

pub(in borrow_tracker::tree_borrows) fn is_relevant( &self, err: TransitionError ) -> bool

Determines whether self is a relevant transition for the error err. self will be a transition that happened to a tag some time before that tag caused the error.

Irrelevant events:

modifications of write permissions when the error is related to read permissions (on failed reads and protected Frozen -> Disabled, ignore Reserved -> Active, Reserved -> Frozen, and Active -> Frozen)

all transitions for attempts to deallocate strongly protected tags

Panics

This function assumes that its arguments apply to the same location and that they were obtained during a normal execution. It will panic otherwise.

all transitions involved in self and err should be increasing (Reserved < Active < Frozen < Disabled);

between self and err the permission should also be increasing, so all permissions inside err should be greater than self.1;

Active and Reserved cannot cause an error due to insufficient permissions, so err cannot be a ChildAccessForbidden(_) of either of them;

err should not be ProtectedDisabled(Disabled), because the protected tag should not have been Disabled in the first place (if this occurs it means we have unprotected tags that become protected)

pub fn endpoint(&self) -> Permission

Endpoint of a transition. Meant only for diagnostics, use applied in non-diagnostics code, which also checks that the starting point matches the current state.

impl PermTransition

fn is_possible(self) -> bool

All transitions created through normal means (using perform_access) should be possible, but the same is not guaranteed by construction of transitions inferred by diagnostics. This checks that a transition reconstructed by diagnostics is indeed one that could happen.

pub fn from(from: Permission, to: Permission) -> Option<Self>

pub fn is_noop(self) -> bool

pub fn applied(self, starting_point: Permission) -> Option<Permission>

Extract result of a transition (checks that the starting point matches).

pub fn started(self) -> Permission

Extract starting point of a transition

pub fn produces_disabled(self) -> bool

Determines if this transition would disable the permission.