Struct miri::borrow_tracker::tree_borrows::perms::Permission

pub struct Permission {
    inner: PermissionPriv,
}

Public interface to the state machine that controls read-write permissions. This is the “private enum” pattern.

Fields
inner: PermissionPriv
Implementations

impl Permission

pub fn short_name(self) -> &'static str

Abbreviated name of the permission (uniformly 3 letters for nice alignment).

impl Permission

pub fn new_active() -> Self

Default initial permission of the root of a new tree.

pub fn new_reserved(ty_is_freeze: bool) -> Self

Default initial permission of a reborrowed mutable reference.

pub fn new_frozen() -> Self

Default initial permission of a reborrowed shared reference

pub fn is_active(self) -> bool

pub fn is_reserved(self) -> bool

pub fn is_frozen(self) -> bool

pub fn perform_access( kind: AccessKind, rel_pos: AccessRelatedness, old_perm: Self, protected: bool ) -> Option<PermTransition>

Apply the transition to the inner PermissionPriv.