Struct miri::borrow_tracker::tree_borrows::tree::TreeVisitorStack

struct TreeVisitorStack<NodeApp, ErrHandler> {
    initial: UniIndex,
    f_propagate: NodeApp,
    err_builder: ErrHandler,
    stack: Vec<(UniIndex, AccessRelatedness)>,
}

Stack of nodes left to explore in a tree traversal.

Fields
initial: UniIndex

Identifier of the original access.

f_propagate: NodeApp

Function to apply to each tag.

err_builder: ErrHandler

Handler to add the required context to diagnostics.

stack: Vec<(UniIndex, AccessRelatedness)>

Mutable state of the visit: the tags left to handle. Every tag pushed should eventually be handled, and the precise order is relevant for diagnostics.

Implementations

impl<NodeApp, InnErr, OutErr, ErrHandler> TreeVisitorStack<NodeApp, ErrHandler>
where NodeApp: Fn(NodeAppArgs<'_>) -> Result<ContinueTraversal, InnErr>, ErrHandler: Fn(ErrHandlerArgs<'_, InnErr>) -> OutErr,

fn exec_and_visit( &mut self, this: &mut TreeVisitor<'_>, idx: UniIndex, exclude: Option<UniIndex>, rel_pos: AccessRelatedness, ) -> Result<(), OutErr>

Apply the function to the current tag, and push its children to the stack of future tags to visit.

fn new(initial: UniIndex, f_propagate: NodeApp, err_builder: ErrHandler) -> Self

fn finish(&mut self, visitor: &mut TreeVisitor<'_>) -> Result<(), OutErr>

Finish the exploration by applying exec_and_visit until the stack is empty.

fn push_and_visit_strict_ancestors( &mut self, visitor: &mut TreeVisitor<'_>, ) -> Result<(), OutErr>

Push all ancestors to the exploration stack in order of nearest ancestor towards the top.