Struct miri::concurrency::data_race::VClockAlloc

pub struct VClockAlloc {
    alloc_ranges: RefCell<RangeMap<MemoryCellClocks>>,
}

Vector clock metadata for a logical memory allocation.

Fields
alloc_ranges: RefCell<RangeMap<MemoryCellClocks>>

Assigning each byte a MemoryCellClocks.

Implementations

impl VClockAlloc

pub fn new_allocation(
global: &GlobalState,
thread_mgr: &ThreadManager<'_, '_>,
len: Size,
kind: MemoryKind<MiriMemoryKind>
) -> VClockAlloc

Create a new data-race detector for newly allocated memory.

fn find_gt_index(l: &VClock, r: &VClock) -> Option<VectorIdx>

fn report_data_race<'tcx>(
global: &GlobalState,
thread_mgr: &ThreadManager<'_, '_>,
range: &MemoryCellClocks,
action: &str,
is_atomic: bool,
ptr_dbg: Pointer<AllocId>
) -> InterpResult<'tcx>

Report a data-race found in the program. This finds the two racing threads and the type of data-race that occurred. This will also return info about the memory location the data-race occurred in.

pub(super) fn race_free_with_atomic(
&self,
range: AllocRange,
global: &GlobalState,
thread_mgr: &ThreadManager<'_, '_>
) -> bool

Detect racing atomic read and writes (not data races) on every byte of the current access range

pub fn read<'tcx>(
&self,
alloc_id: AllocId,
range: AllocRange,
global: &GlobalState,
thread_mgr: &ThreadManager<'_, '_>
) -> InterpResult<'tcx>

Detect data-races for an unsynchronized read operation, will not perform data-race detection if race_detecting() is false, either due to no threads being created or if it is temporarily disabled during a racy read or write operation for which data-race detection is handled separately, for example atomic read operations.

fn unique_access<'tcx>(
&mut self,
alloc_id: AllocId,
range: AllocRange,
write_type: WriteType,
global: &mut GlobalState,
thread_mgr: &ThreadManager<'_, '_>
) -> InterpResult<'tcx>

pub fn write<'tcx>(
&mut self,
alloc_id: AllocId,
range: AllocRange,
global: &mut GlobalState,
thread_mgr: &ThreadManager<'_, '_>
) -> InterpResult<'tcx>

Detect data-races for an unsynchronized write operation, will not perform data-race threads if race_detecting() is false, either due to no threads being created or if it is temporarily disabled during a racy read or write operation

pub fn deallocate<'tcx>(
&mut self,
alloc_id: AllocId,
range: AllocRange,
global: &mut GlobalState,
thread_mgr: &ThreadManager<'_, '_>
) -> InterpResult<'tcx>

Detect data-races for an unsynchronized deallocate operation, will not perform data-race threads if race_detecting() is false, either due to no threads being created or if it is temporarily disabled during a racy read or write operation