Trait miri::concurrency::sync::EvalContextExtPriv

trait EvalContextExtPriv<'mir, 'tcx: 'mir>: MiriEvalContextExt<'mir, 'tcx> {
    fn rwlock_dequeue_and_lock_reader(&mut self, id: RwLockId) -> bool { ... }
    fn rwlock_dequeue_and_lock_writer(&mut self, id: RwLockId) -> bool { ... }
    fn mutex_dequeue_and_lock(&mut self, id: MutexId) -> bool { ... }
}

Provided Methods

fn rwlock_dequeue_and_lock_reader(&mut self, id: RwLockId) -> bool

Take a reader out of the queue waiting for the lock. Returns true if some thread got the rwlock.

fn rwlock_dequeue_and_lock_writer(&mut self, id: RwLockId) -> bool

Take the writer out of the queue waiting for the lock. Returns true if some thread got the rwlock.

fn mutex_dequeue_and_lock(&mut self, id: MutexId) -> bool

Take a thread out of the queue waiting for the mutex, and lock the mutex for it. Returns true if some thread has the mutex now.