Module miri::concurrency::thread

Implements threads.

Structs

Thread

A thread.

ThreadId

A thread identifier.

ThreadManager

A set of threads.

TimeoutCallbackInfo 🔒

Callbacks are used to implement timeouts. For example, waiting on a conditional variable with a timeout creates a callback that is called after the specified time and unblocks the thread. If another thread signals on the conditional variable, the signal handler deletes the callback.

Enums

SchedulingAction 🔒

ThreadJoinStatus 🔒

The join status of a thread.

ThreadState

The state of a thread.

Time

A specific moment in time.

Traits

EvalContextExt

EvalContextPrivExt 🔒

MachineCallback

Trait for callbacks that can be executed when some event happens, such as after a timeout.

Type Definitions

StackEmptyCallback

TimeoutCallback 🔒