Type Alias miri::concurrency::weak_memory::AllocState

pub type AllocState = StoreBufferAlloc;

Aliased Type

struct AllocState {
    store_buffers: RefCell<RangeObjectMap<StoreBuffer>>,
}

Fields
store_buffers: RefCell<RangeObjectMap<StoreBuffer>>

Store buffer of each atomic object in this allocation

Implementations

impl StoreBufferAlloc

pub fn new_allocation() -> Self

fn is_overlapping(&self, range: AllocRange) -> bool

Checks if the range imperfectly overlaps with existing buffers Used to determine if mixed-size atomic accesses

pub fn memory_accessed(&self, range: AllocRange, global: &DataRaceState)

When a non-atomic access happens on a location that has been atomically accessed before without data race, we can determine that the non-atomic access fully happens after all the prior atomic accesses so the location no longer needs to exhibit any weak memory behaviours until further atomic accesses.

fn get_or_create_store_buffer<'tcx>( &self, range: AllocRange, init: Scalar<Provenance> ) -> InterpResult<'tcx, Ref<'_, StoreBuffer>>

Gets a store buffer associated with an atomic object in this allocation, or creates one with the specified initial value if no atomic object exists yet.

fn get_or_create_store_buffer_mut<'tcx>( &mut self, range: AllocRange, init: Scalar<Provenance> ) -> InterpResult<'tcx, &mut StoreBuffer>

Gets a mutable store buffer associated with an atomic object in this allocation