Function miri::diagnostics::report_msg

fn report_msg<'mir, 'tcx>(
    ecx: &InterpCx<'mir, 'tcx, Evaluator<'mir, 'tcx>>,
    diag_level: DiagLevel,
    title: &str,
    span_msg: Vec<String>,
    notes: Vec<(Option<SpanData>, String)>,
    helps: Vec<(Option<SpanData>, String)>,
    stacktrace: &[FrameInfo<'tcx>]
)

Report an error or note (depending on the error argument) with the given stacktrace. Also emits a full stacktrace of the interpreter stack. We want to present a multi-line span message for some errors. Diagnostics do not support this directly, so we pass the lines as a Vec<String> and display each line after the first with an additional span_label or note call.