Enum miri::MiriMemoryKind

pub enum MiriMemoryKind {
    Rust,
    Miri,
    C,
    WinHeap,
    WinLocal,
    Machine,
    Runtime,
    Global,
    ExternStatic,
    Tls,
    Mmap,
}

Extra memory kinds

Variants

Rust

__rust_alloc memory.

Miri

miri_alloc memory.

C

malloc memory.

WinHeap

Windows HeapAlloc memory.

WinLocal

Windows “local” memory (to be freed with LocalFree)

Machine

Memory for args, errno, and other parts of the machine-managed environment. This memory may leak.

Runtime

Memory allocated by the runtime (e.g. env vars). Separate from Machine because we clean it up and leak-check it.

Global

Globals copied from tcx. This memory may leak.

ExternStatic

Memory for extern statics. This memory may leak.

Tls

Memory for thread-local statics. This memory may leak.

Mmap

Memory mapped directly by the program