Function miri::intern_const_alloc_recursive

pub fn intern_const_alloc_recursive<'mir, 'tcx, M>(
    ecx: &mut InterpCx<'mir, 'tcx, M>,
    intern_kind: InternKind,
    ret: &MPlaceTy<'tcx, AllocId>
) -> Result<(), ErrorGuaranteed>where
    'tcx: 'mir,
    M: CompileTimeMachine<'mir, 'tcx, MemoryKind>,

Intern ret and everything it references.

This cannot raise an interpreter error. Doing so is left to validation, which tracks where in the value we are and thus can show much better error messages. Any errors here would anyway be turned into const_err lints, whereas validation failures are hard errors.