Module miri::helpers

Modules

convert

Structs

CurrentSpan

A CurrentSpan should be created infrequently (ideally once) per interpreter step. It does nothing on creation, but when CurrentSpan::get is called, searches the current stack for the topmost frame which corresponds to a local crate, and returns the current span in that frame. The result of that search is cached so that later calls are approximately free.

Constants

UNIX_IO_ERROR_TABLE 🔒

Traits

EvalContextExt

Functions

check_arg_count

Check that the number of args is what we expect.

get_local_crates

Retrieve the list of local crates that should have been passed by cargo-miri in MIRI_LOCAL_CRATES and turn them into CrateNums.

isolation_abort_error

target_os_is_unix

Helper function used inside the shims of foreign functions to check that target_os is a supported UNIX OS.

try_resolve_did 🔒

Gets an instance for a path.