Enum miri::interpret::UndefinedBehaviorInfo

pub enum UndefinedBehaviorInfo {

    Ub(String),
    Unreachable,
    BoundsCheckFailed {
        len: u64,
        index: u64,
    },
    DivisionByZero,
    RemainderByZero,
    DivisionOverflow,
    RemainderOverflow,
    PointerArithOverflow,
    InvalidMeta(&'static str),
    UnterminatedCString(Pointer<AllocId>),
    PointerUseAfterFree(AllocId),
    PointerOutOfBounds {
        alloc_id: AllocId,
        alloc_size: Size,
        ptr_offset: i64,
        ptr_size: Size,
        msg: CheckInAllocMsg,
    },
    DanglingIntPointer(u64, CheckInAllocMsg),
    AlignmentCheckFailed {
        required: Align,
        has: Align,
    },
    WriteToReadOnly(AllocId),
    DerefFunctionPointer(AllocId),
    DerefVTablePointer(AllocId),
    ValidationFailure {
        path: Option<String>,
        msg: String,
    },
    InvalidBool(u8),
    InvalidChar(u32),
    InvalidTag(Scalar<AllocId>),
    InvalidFunctionPointer(Pointer<AllocId>),
    InvalidVTablePointer(Pointer<AllocId>),
    InvalidStr(Utf8Error),
    InvalidUninitBytes(Option<(AllocId, UninitBytesAccess)>),
    DeadLocal,
    ScalarSizeMismatch(ScalarSizeMismatch),
    UninhabitedEnumVariantWritten,

}

Error information for when the program caused Undefined Behavior.

Variants

Ub(String)

Free-form case. Only for errors that are never caught!

Unreachable

Unreachable code was executed.

BoundsCheckFailed

Fields

len: u64

index: u64

A slice/array index projection went out-of-bounds.

DivisionByZero

Something was divided by 0 (x / 0).

RemainderByZero

Something was “remainded” by 0 (x % 0).

DivisionOverflow

Signed division overflowed (INT_MIN / -1).

RemainderOverflow

Signed remainder overflowed (INT_MIN % -1).

PointerArithOverflow

Overflowing inbounds pointer arithmetic.

InvalidMeta(&'static str)

Invalid metadata in a wide pointer (using str to avoid allocations).

UnterminatedCString(Pointer<AllocId>)

Reading a C string that does not end within its allocation.

PointerUseAfterFree(AllocId)

Dereferencing a dangling pointer after it got freed.

PointerOutOfBounds

Fields

alloc_id: AllocId

alloc_size: Size

ptr_offset: i64

ptr_size: Size

msg: CheckInAllocMsg

Used a pointer outside the bounds it is valid for. (If ptr_size > 0, determines the size of the memory range that was expected to be in-bounds.)

DanglingIntPointer(u64, CheckInAllocMsg)

Using an integer as a pointer in the wrong way.

AlignmentCheckFailed

Fields

required: Align

has: Align

Used a pointer with bad alignment.

WriteToReadOnly(AllocId)

Writing to read-only memory.

DerefFunctionPointer(AllocId)

DerefVTablePointer(AllocId)

ValidationFailure

Fields

path: Option<String>

The “path” to the value in question, e.g. .0[5].field for a struct field in the 6th element of an array that is the first element of a tuple.

msg: String

The value validity check found a problem. Should only be thrown by validity.rs and always point out which part of the value is the problem.

InvalidBool(u8)

Using a non-boolean u8 as bool.

InvalidChar(u32)

Using a non-character u32 as character.

InvalidTag(Scalar<AllocId>)

The tag of an enum does not encode an actual discriminant.

InvalidFunctionPointer(Pointer<AllocId>)

Using a pointer-not-to-a-function as function pointer.

InvalidVTablePointer(Pointer<AllocId>)

Using a pointer-not-to-a-vtable as vtable pointer.

InvalidStr(Utf8Error)

Using a string that is not valid UTF-8,

InvalidUninitBytes(Option<(AllocId, UninitBytesAccess)>)

Using uninitialized data where it is not allowed.

DeadLocal

Working with a local that is not currently live.

ScalarSizeMismatch(ScalarSizeMismatch)

Data size is not equal to target size.

UninhabitedEnumVariantWritten

A discriminant of an uninhabited enum variant is written.