Struct miri::interpret::Frame

pub struct Frame<'mir, 'tcx, Prov = AllocId, Extra = ()>where
    Prov: Provenance,{
    pub body: &'mir Body<'tcx>,
    pub instance: Instance<'tcx>,
    pub extra: Extra,
    pub return_to_block: StackPopCleanup,
    pub return_place: PlaceTy<'tcx, Prov>,
    pub locals: IndexVec<Local, LocalState<'tcx, Prov>>,
    pub loc: Either<Location, Span>,
    /* private fields */
}

A stack frame.

Fields
body: &'mir Body<'tcx>

The MIR for the function called on this frame.

instance: Instance<'tcx>

The def_id and substs of the current function.

extra: Extra

Extra data for the machine.

return_to_block: StackPopCleanup

Work to perform when returning from this function.

return_place: PlaceTy<'tcx, Prov>

The location where the result of the current stack frame should be written to, and its layout in the caller.

locals: IndexVec<Local, LocalState<'tcx, Prov>>

The list of locals for this stack frame, stored in order as [return_ptr, arguments..., variables..., temporaries...]. The locals are stored as Option<Value>s. None represents a local that is currently dead, while a live local can either directly contain Scalar or refer to some part of an Allocation.

Do not access this directly; always go through the machine hook!

loc: Either<Location, Span>

If this is Right, we are not currently executing any particular statement in this frame (can happen e.g. during frame initialization, and during unwinding on frames without cleanup code).

Needs to be public because ConstProp does unspeakable things to it.