Trait miri::interpret::ValueVisitor

pub trait ValueVisitor<'mir, 'tcx, M>: Sizedwhere
    'tcx: 'mir,
    M: Machine<'mir, 'tcx>,{
    type V: Value<'mir, 'tcx, M>;

    fn ecx(&self) -> &InterpCx<'mir, 'tcx, M>;

    fn read_discriminant(
        &mut self,
        op: &OpTy<'tcx, <M as Machine<'mir, 'tcx>>::Provenance>
    ) -> Result<VariantIdx, InterpErrorInfo<'tcx>> { ... }
    fn visit_value(&mut self, v: &Self::V) -> Result<(), InterpErrorInfo<'tcx>> { ... }
    fn visit_union(
        &mut self,
        _v: &Self::V,
        _fields: NonZeroUsize
    ) -> Result<(), InterpErrorInfo<'tcx>> { ... }
    fn visit_box(&mut self, _v: &Self::V) -> Result<(), InterpErrorInfo<'tcx>> { ... }
    fn visit_aggregate(
        &mut self,
        v: &Self::V,
        fields: impl Iterator<Item = Result<Self::V, InterpErrorInfo<'tcx>>>
    ) -> Result<(), InterpErrorInfo<'tcx>> { ... }
    fn visit_field(
        &mut self,
        _old_val: &Self::V,
        _field: usize,
        new_val: &Self::V
    ) -> Result<(), InterpErrorInfo<'tcx>> { ... }
    fn visit_variant(
        &mut self,
        _old_val: &Self::V,
        _variant: VariantIdx,
        new_val: &Self::V
    ) -> Result<(), InterpErrorInfo<'tcx>> { ... }
    fn walk_aggregate(
        &mut self,
        v: &Self::V,
        fields: impl Iterator<Item = Result<Self::V, InterpErrorInfo<'tcx>>>
    ) -> Result<(), InterpErrorInfo<'tcx>> { ... }
    fn walk_value(&mut self, v: &Self::V) -> Result<(), InterpErrorInfo<'tcx>> { ... }
}

How to traverse a value and what to do when we are at the leaves.

Required Associated Types

type V: Value<'mir, 'tcx, M>

Required Methods

fn ecx(&self) -> &InterpCx<'mir, 'tcx, M>

The visitor must have an InterpCx in it.

Provided Methods

fn read_discriminant(
&mut self,
op: &OpTy<'tcx, <M as Machine<'mir, 'tcx>>::Provenance>
) -> Result<VariantIdx, InterpErrorInfo<'tcx>>

read_discriminant can be hooked for better error messages.

fn visit_value(&mut self, v: &Self::V) -> Result<(), InterpErrorInfo<'tcx>>

Visits the given value, dispatching as appropriate to more specialized visitors.

fn visit_union(
&mut self,
_v: &Self::V,
_fields: NonZeroUsize
) -> Result<(), InterpErrorInfo<'tcx>>

Visits the given value as a union. No automatic recursion can happen here.

fn visit_box(&mut self, _v: &Self::V) -> Result<(), InterpErrorInfo<'tcx>>

Visits the given value as the pointer of a Box. There is nothing to recurse into. The type of v will be a raw pointer, but this is a field of Box<T> and the pointee type is the actual T.

fn visit_aggregate(
&mut self,
v: &Self::V,
fields: impl Iterator<Item = Result<Self::V, InterpErrorInfo<'tcx>>>
) -> Result<(), InterpErrorInfo<'tcx>>

Visits this value as an aggregate, you are getting an iterator yielding all the fields (still in an InterpResult, you have to do error handling yourself). Recurses into the fields.

fn visit_field(
&mut self,
_old_val: &Self::V,
_field: usize,
new_val: &Self::V
) -> Result<(), InterpErrorInfo<'tcx>>

Called each time we recurse down to a field of a “product-like” aggregate (structs, tuples, arrays and the like, but not enums), passing in old (outer) and new (inner) value. This gives the visitor the chance to track the stack of nested fields that we are descending through.

fn visit_variant(
&mut self,
_old_val: &Self::V,
_variant: VariantIdx,
new_val: &Self::V
) -> Result<(), InterpErrorInfo<'tcx>>

Called when recursing into an enum variant. This gives the visitor the chance to track the stack of nested fields that we are descending through.

fn walk_aggregate(
&mut self,
v: &Self::V,
fields: impl Iterator<Item = Result<Self::V, InterpErrorInfo<'tcx>>>
) -> Result<(), InterpErrorInfo<'tcx>>

fn walk_value(&mut self, v: &Self::V) -> Result<(), InterpErrorInfo<'tcx>>