miri

Module machine

Global machine state as well as implementation of the interpreter engine Machine trait.

Structs

AllocExtra

Extra per-allocation data

FrameExtra

Extra data stored with each stack frame

MiriMachine

The machine itself.

PrimitiveLayouts

Precomputed layouts of primitive types

Enums

MiriMemoryKind

Extra memory kinds

Provenance

Pointer provenance.

ProvenanceExtra

The “extra” information a pointer has over a regular AllocId.

Constants

ADDRS_PER_ANON_GLOBAL 🔒

Each anonymous global (constant, vtable, function pointer, …) has multiple addresses, but only this many. Since const allocations are never deallocated, choosing a new AllocId and thus base address for each evaluation would produce unbounded memory usage.

SIGRTMAX

Last real-time signal. signal(7) says it must be between 32 and 64 and specifies SIGRTMAX - SIGRTMIN >= 8 (which is the value of _POSIX_RTSIG_MAX)

SIGRTMIN

First real-time signal. signal(7) says this must be between 32 and 64 and specifies 34 or 35 as typical values.

Traits

MachineCallback

Trait for callbacks handling asynchronous machine operations.

MiriInterpCxExt

A little trait that’s useful to be inherited by extension traits.

Type Aliases

DynMachineCallback

Type alias for boxed machine callbacks with generic argument type.

MemoryKind

MiriInterpCx

A rustc InterpCx for Miri.