Struct miri::machine::AllocExtra

pub struct AllocExtra<'tcx> {
    pub borrow_tracker: Option<AllocState>,
    pub data_race: Option<VClockAlloc>,
    pub weak_memory: Option<StoreBufferAlloc>,
    pub backtrace: Option<Vec<FrameInfo<'tcx>>>,
}

Extra per-allocation data

Fields
borrow_tracker: Option<AllocState>

Global state of the borrow tracker, if enabled.

data_race: Option<VClockAlloc>

Data race detection via the use of a vector-clock. This is only added if it is enabled.

weak_memory: Option<StoreBufferAlloc>

Weak memory emulation via the use of store buffers. This is only added if it is enabled.

backtrace: Option<Vec<FrameInfo<'tcx>>>

A backtrace to where this allocation was allocated. As this is recorded for leak reports, it only exists if this allocation is leakable. The backtrace is not pruned yet; that should be done before printing it.

Implementations

impl AllocExtra<'_>

pub fn borrow_tracker_sb(&self) -> &RefCell<Stacks>

pub fn borrow_tracker_sb_mut(&mut self) -> &mut RefCell<Stacks>

pub fn borrow_tracker_tb(&self) -> &RefCell<Tree>