Struct miri::machine::Evaluator

pub struct Evaluator<'mir, 'tcx> {

    pub stacked_borrows: Option<RefCell<GlobalStateInner>>,
    pub data_race: Option<GlobalState>,
    pub intptrcast: RefCell<GlobalStateInner>,
    pub(crate) env_vars: EnvVars<'tcx>,
    pub(crate) argc: Option<MemPlace<Provenance>>,
    pub(crate) argv: Option<MemPlace<Provenance>>,
    pub(crate) cmd_line: Option<MemPlace<Provenance>>,
    pub(crate) tls: TlsData<'tcx>,
    pub(crate) isolated_op: IsolatedOp,
    pub(crate) validate: bool,
    pub(crate) enforce_abi: bool,
    pub(crate) file_handler: FileHandler,
    pub(crate) dir_handler: DirHandler,
    pub(crate) time_anchor: Instant,
    pub(crate) threads: ThreadManager<'mir, 'tcx>,
    pub(crate) layouts: PrimitiveLayouts<'tcx>,
    pub(crate) static_roots: Vec<AllocId>,
    profiler: Option<Profiler>,
    string_cache: FxHashMap<String, StringId>,
    pub(crate) exported_symbols_cache: FxHashMap<Symbol, Option<Instance<'tcx>>>,
    pub(crate) panic_on_unsupported: bool,
    pub(crate) backtrace_style: BacktraceStyle,
    pub(crate) local_crates: Vec<CrateNum>,
    extern_statics: FxHashMap<Symbol, Pointer<Provenance>>,
    pub(crate) rng: RefCell<StdRng>,
    tracked_alloc_ids: FxHashSet<AllocId>,
    pub(crate) check_alignment: AlignmentCheck,
    pub(crate) cmpxchg_weak_failure_rate: f64,
    pub(crate) mute_stdout_stderr: bool,
    pub(crate) weak_memory: bool,
    pub(crate) preemption_rate: f64,
    pub(crate) report_progress: Option<u32>,
    pub(crate) basic_block_count: u64,
    pub external_so_lib: Option<(Library, PathBuf)>,

}

The machine itself.

Fields
stacked_borrows: Option<RefCell<GlobalStateInner>>data_race: Option<GlobalState>intptrcast: RefCell<GlobalStateInner>env_vars: EnvVars<'tcx>

Environment variables set by setenv. Miri does not expose env vars from the host to the emulated program.

argc: Option<MemPlace<Provenance>>

Program arguments (Option because we can only initialize them after creating the ecx). These are pointers to argc/argv because macOS. We also need the full command line as one string because of Windows.

argv: Option<MemPlace<Provenance>>cmd_line: Option<MemPlace<Provenance>>tls: TlsData<'tcx>

TLS state.

isolated_op: IsolatedOp

What should Miri do when an op requires communicating with the host, such as accessing host env vars, random number generation, and file system access.

validate: bool

Whether to enforce the validity invariant.

enforce_abi: bool

Whether to enforce ABI of function calls.

file_handler: FileHandler

The table of file descriptors.

dir_handler: DirHandler

The table of directory descriptors.

time_anchor: Instant

The “time anchor” for this machine’s monotone clock (for Instant simulation).

threads: ThreadManager<'mir, 'tcx>

The set of threads.

layouts: PrimitiveLayouts<'tcx>

Precomputed TyLayouts for primitive data types that are commonly used inside Miri.

static_roots: Vec<AllocId>

Allocations that are considered roots of static memory (that may leak).

profiler: Option<Profiler>

The measureme profiler used to record timing information about the emulated program.

string_cache: FxHashMap<String, StringId>

Used with profiler to cache the StringIds for event names uesd with measureme.

exported_symbols_cache: FxHashMap<Symbol, Option<Instance<'tcx>>>

Cache of Instance exported under the given Symbol name. None means no Instance exported under the given name is found.

panic_on_unsupported: bool

Whether to raise a panic in the context of the evaluated process when unsupported functionality is encountered. If false, an error is propagated in the Miri application context instead (default behavior)

backtrace_style: BacktraceStyle

Equivalent setting as RUST_BACKTRACE on encountering an error.

local_crates: Vec<CrateNum>

Crates which are considered local for the purposes of error reporting.

extern_statics: FxHashMap<Symbol, Pointer<Provenance>>

Mapping extern static names to their base pointer.

rng: RefCell<StdRng>

The random number generator used for resolving non-determinism. Needs to be queried by ptr_to_int, hence needs interior mutability.

tracked_alloc_ids: FxHashSet<AllocId>

The allocation IDs to report when they are being allocated (helps for debugging memory leaks and use after free bugs).

check_alignment: AlignmentCheck

Controls whether alignment of memory accesses is being checked.

cmpxchg_weak_failure_rate: f64

Failure rate of compare_exchange_weak, between 0.0 and 1.0

mute_stdout_stderr: bool

Corresponds to -Zmiri-mute-stdout-stderr and doesn’t write the output but acts as if it succeeded.

weak_memory: bool

Whether weak memory emulation is enabled

preemption_rate: f64

The probability of the active thread being preempted at the end of each basic block.

report_progress: Option<u32>

If Some, we will report the current stack every N basic blocks.

basic_block_count: u64external_so_lib: Option<(Library, PathBuf)>

Handle of the optional shared object file for external functions.

Implementations

impl<'mir, 'tcx> Evaluator<'mir, 'tcx>

pub fn current_span(&self, tcx: TyCtxt<'tcx>) -> CurrentSpan<'_, 'mir, 'tcx>

impl<'mir, 'tcx> Evaluator<'mir, 'tcx>

pub(crate) fn new(
config: &MiriConfig,
layout_cx: LayoutCx<'tcx, TyCtxt<'tcx>>
) -> Self

pub(crate) fn late_init(
this: &mut MiriEvalContext<'mir, 'tcx>,
config: &MiriConfig
) -> InterpResult<'tcx>

fn add_extern_static(
this: &mut MiriEvalContext<'mir, 'tcx>,
name: &str,
ptr: Pointer<Option<Provenance>>
)

fn alloc_extern_static(
this: &mut MiriEvalContext<'mir, 'tcx>,
name: &str,
val: ImmTy<'tcx, Provenance>
) -> InterpResult<'tcx>

fn init_extern_statics(
this: &mut MiriEvalContext<'mir, 'tcx>
) -> InterpResult<'tcx>

Sets up the “extern statics” for this machine.

pub(crate) fn communicate(&self) -> bool

pub(crate) fn is_local(&self, frame: &FrameInfo<'_>) -> bool

Check whether the stack frame that this FrameInfo refers to is part of a local crate.