Struct miri::machine::MiriMachine

pub struct MiriMachine<'mir, 'tcx> {

    pub tcx: TyCtxt<'tcx>,
    pub borrow_tracker: Option<RefCell<GlobalStateInner>>,
    pub data_race: Option<GlobalState>,
    pub intptrcast: RefCell<GlobalStateInner>,
    pub(crate) env_vars: EnvVars<'tcx>,
    pub(crate) main_fn_ret_place: Option<MemPlace<Provenance>>,
    pub(crate) argc: Option<MemPlace<Provenance>>,
    pub(crate) argv: Option<MemPlace<Provenance>>,
    pub(crate) cmd_line: Option<MemPlace<Provenance>>,
    pub(crate) tls: TlsData<'tcx>,
    pub(crate) isolated_op: IsolatedOp,
    pub(crate) validate: bool,
    pub(crate) enforce_abi: bool,
    pub(crate) file_handler: FileHandler,
    pub(crate) dir_handler: DirHandler,
    pub(crate) clock: Clock,
    pub(crate) threads: ThreadManager<'mir, 'tcx>,
    pub(crate) layouts: PrimitiveLayouts<'tcx>,
    pub(crate) static_roots: Vec<AllocId>,
    profiler: Option<Profiler>,
    string_cache: FxHashMap<String, StringId>,
    pub(crate) exported_symbols_cache: FxHashMap<Symbol, Option<Instance<'tcx>>>,
    pub(crate) panic_on_unsupported: bool,
    pub(crate) backtrace_style: BacktraceStyle,
    pub(crate) local_crates: Vec<CrateNum>,
    extern_statics: FxHashMap<Symbol, Pointer<Provenance>>,
    pub(crate) rng: RefCell<StdRng>,
    tracked_alloc_ids: FxHashSet<AllocId>,
    pub(crate) check_alignment: AlignmentCheck,
    pub(crate) cmpxchg_weak_failure_rate: f64,
    pub(crate) mute_stdout_stderr: bool,
    pub(crate) weak_memory: bool,
    pub(crate) preemption_rate: f64,
    pub(crate) report_progress: Option<u32>,
    pub(crate) basic_block_count: u64,
    pub external_so_lib: Option<(Library, PathBuf)>,
    pub(crate) gc_interval: u32,
    pub(crate) since_gc: u32,
    pub(crate) num_cpus: u32,
    pub(crate) page_size: u64,
    pub(crate) stack_addr: u64,
    pub(crate) stack_size: u64,
    pub(crate) collect_leak_backtraces: bool,
    pub(crate) allocation_spans: RefCell<FxHashMap<AllocId, (Span, Option<Span>)>>,

}

The machine itself.

If you add anything here that stores machine values, remember to update visit_all_machine_values!

Fields
tcx: TyCtxt<'tcx>borrow_tracker: Option<RefCell<GlobalStateInner>>

Global data for borrow tracking.

data_race: Option<GlobalState>

Data race detector global data.

intptrcast: RefCell<GlobalStateInner>

Ptr-int-cast module global data.

env_vars: EnvVars<'tcx>

Environment variables set by setenv. Miri does not expose env vars from the host to the emulated program.

main_fn_ret_place: Option<MemPlace<Provenance>>

Return place of the main function.

argc: Option<MemPlace<Provenance>>

Program arguments (Option because we can only initialize them after creating the ecx). These are pointers to argc/argv because macOS. We also need the full command line as one string because of Windows.

argv: Option<MemPlace<Provenance>>cmd_line: Option<MemPlace<Provenance>>tls: TlsData<'tcx>

TLS state.

isolated_op: IsolatedOp

What should Miri do when an op requires communicating with the host, such as accessing host env vars, random number generation, and file system access.

validate: bool

Whether to enforce the validity invariant.

enforce_abi: bool

Whether to enforce ABI of function calls.

file_handler: FileHandler

The table of file descriptors.

dir_handler: DirHandler

The table of directory descriptors.

clock: Clock

This machine’s monotone clock.

threads: ThreadManager<'mir, 'tcx>

The set of threads.

layouts: PrimitiveLayouts<'tcx>

Precomputed TyLayouts for primitive data types that are commonly used inside Miri.

static_roots: Vec<AllocId>

Allocations that are considered roots of static memory (that may leak).

profiler: Option<Profiler>

The measureme profiler used to record timing information about the emulated program.

string_cache: FxHashMap<String, StringId>

Used with profiler to cache the StringIds for event names used with measureme.

exported_symbols_cache: FxHashMap<Symbol, Option<Instance<'tcx>>>

Cache of Instance exported under the given Symbol name. None means no Instance exported under the given name is found.

panic_on_unsupported: bool

Whether to raise a panic in the context of the evaluated process when unsupported functionality is encountered. If false, an error is propagated in the Miri application context instead (default behavior)

backtrace_style: BacktraceStyle

Equivalent setting as RUST_BACKTRACE on encountering an error.

local_crates: Vec<CrateNum>

Crates which are considered local for the purposes of error reporting.

extern_statics: FxHashMap<Symbol, Pointer<Provenance>>

Mapping extern static names to their base pointer.

rng: RefCell<StdRng>

The random number generator used for resolving non-determinism. Needs to be queried by ptr_to_int, hence needs interior mutability.

tracked_alloc_ids: FxHashSet<AllocId>

The allocation IDs to report when they are being allocated (helps for debugging memory leaks and use after free bugs).

check_alignment: AlignmentCheck

Controls whether alignment of memory accesses is being checked.

cmpxchg_weak_failure_rate: f64

Failure rate of compare_exchange_weak, between 0.0 and 1.0

mute_stdout_stderr: bool

Corresponds to -Zmiri-mute-stdout-stderr and doesn’t write the output but acts as if it succeeded.

weak_memory: bool

Whether weak memory emulation is enabled

preemption_rate: f64

The probability of the active thread being preempted at the end of each basic block.

report_progress: Option<u32>

If Some, we will report the current stack every N basic blocks.

basic_block_count: u64external_so_lib: Option<(Library, PathBuf)>

Handle of the optional shared object file for external functions.

gc_interval: u32

Run a garbage collector for BorTags every N basic blocks.

since_gc: u32

The number of blocks that passed since the last BorTag GC pass.

num_cpus: u32

The number of CPUs to be reported by miri.

page_size: u64

Determines Miri’s page size and associated values

stack_addr: u64stack_size: u64collect_leak_backtraces: bool

Whether to collect a backtrace when each allocation is created, just in case it leaks.

allocation_spans: RefCell<FxHashMap<AllocId, (Span, Option<Span>)>>

The spans we will use to report where an allocation was created and deallocated in diagnostics.

Implementations

impl<'mir, 'tcx> MiriMachine<'mir, 'tcx>

pub fn emit_diagnostic(&self, e: NonHaltingDiagnostic)

impl<'mir, 'tcx> MiriMachine<'mir, 'tcx>

pub fn current_span(&self) -> Span

Get the current span in the topmost function which is workspace-local and not #[track_caller]. This function is backed by a cache, and can be assumed to be very fast. It will work even when the stack is empty.

pub fn caller_span(&self) -> Span

Returns the span of the caller of the current operation, again walking down the stack to find the closest frame in a local crate, if the caller of the current operation is not in a local crate. This is useful when we are processing something which occurs on function-entry and we want to point at the call to the function, not the function definition generally.

fn stack(&self) -> &[Frame<'mir, 'tcx, Provenance, FrameExtra<'tcx>>]

fn top_user_relevant_frame(&self) -> Option<usize>

pub fn is_user_relevant(&self, frame: &Frame<'mir, 'tcx, Provenance>) -> bool

This is the source of truth for the is_user_relevant flag in our FrameExtra.

impl<'mir, 'tcx> MiriMachine<'mir, 'tcx>

pub(crate) fn new( config: &MiriConfig, layout_cx: LayoutCx<'tcx, TyCtxt<'tcx>> ) -> Self

pub(crate) fn late_init( this: &mut MiriInterpCx<'mir, 'tcx>, config: &MiriConfig, on_main_stack_empty: StackEmptyCallback<'mir, 'tcx> ) -> InterpResult<'tcx>

fn add_extern_static( this: &mut MiriInterpCx<'mir, 'tcx>, name: &str, ptr: Pointer<Option<Provenance>> )

fn alloc_extern_static( this: &mut MiriInterpCx<'mir, 'tcx>, name: &str, val: ImmTy<'tcx, Provenance> ) -> InterpResult<'tcx>

fn init_extern_statics( this: &mut MiriInterpCx<'mir, 'tcx> ) -> InterpResult<'tcx>

Sets up the “extern statics” for this machine.

pub(crate) fn communicate(&self) -> bool

pub(crate) fn is_local(&self, frame: &FrameInfo<'_>) -> bool

Check whether the stack frame that this FrameInfo refers to is part of a local crate.

pub(crate) fn handle_abnormal_termination(&mut self)

Called when the interpreter is going to shut down abnormally, such as due to a Ctrl-C.

pub(crate) fn round_up_to_multiple_of_page_size( &self, length: u64 ) -> Option<u64>

pub(crate) fn page_align(&self) -> Align

pub(crate) fn allocated_span(&self, alloc_id: AllocId) -> Option<SpanData>

pub(crate) fn deallocated_span(&self, alloc_id: AllocId) -> Option<SpanData>