miri

Macro enter_trace_span

macro_rules! enter_trace_span {
    ($($tt:tt)*) => { ... };
}

Enters a tracing::info_span only if the “tracing” feature is enabled, otherwise does nothing. This calls rustc_const_eval::enter_trace_span with MiriMachine as the first argument, which will in turn call MiriMachine::enter_trace_span, which takes care of determining at compile time whether to trace or not (and supposedly the call is compiled out if tracing is disabled). Look at rustc_const_eval::enter_trace_span for complete documentation, examples and tips.