miri

Module math

Traits

IeeeExt

Extend functionality of rustc_apfloat softfloats

Functions

apply_random_float_error 🔒

Disturbes a floating-point result by a relative error in the range (-2^scale, 2^scale).

apply_random_float_error_to_imm 🔒

Applies a random 16ULP floating point error to val and returns the new value. Will fail if val is not a floating point number.

apply_random_float_error_ulp 🔒

apply_random_float_error gives instructions to apply a 2^N ULP error. This function implements these instructions such that applying a 2^N ULP error is less error prone. So for a 2^N ULP error, you would pass N as the ulp_exponent argument.

sqrt 🔒