miri::math

Function apply_random_float_error_ulp

pub(crate) fn apply_random_float_error_ulp<F: Float>(
    ecx: &mut MiriInterpCx<'_>,
    val: F,
    ulp_exponent: u32,
) -> F

apply_random_float_error gives instructions to apply a 2^N ULP error. This function implements these instructions such that applying a 2^N ULP error is less error prone. So for a 2^N ULP error, you would pass N as the ulp_exponent argument.