miri

Module native_lib

Functions

init_sv⚠

Initialises the supervisor process. If this function errors, then the supervisor process could not be created successfully; else, the caller is now the child process and can communicate via do_ffi, receiving back events at the end.

register_retcode_sv

Instruct the supervisor process to return a particular code. Useful if for whatever reason this code fails to be intercepted normally.