miri::shims

Module files

Structs

FdId

A unique id for file descriptions. While we could use the address, considering that is definitely unique, the address would expose interpreter internal state when used for sorting things. So instead we generate a unique id per file description is the name for all duplicates and is never reused.

FdIdWith 🔒

FdTable

The file descriptor table

FileDescriptionRef

A refcounted pointer to a file description, also tracking the globally unique ID of this file description.

FileHandle

NullOutput

Like /dev/null

WeakFileDescriptionRef

Holds a weak reference to the actual file description.

Traits

EvalContextExt

FileDescription

Represents an open file description.

FileDescriptionExt

A helper trait to indirectly allow downcasting on Rc<FdIdWith<dyn _>>. Ideally we’d just add a FdIdWith<Self>: Any bound to the FileDescription trait, but that does not allow upcasting.

Type Aliases

DynFileDescriptionRef

FdNum

Internal type of a file-descriptor - this is what FdTable expects