miri::shims::files

Trait EvalContextExt

pub trait EvalContextExt<'tcx>: MiriInterpCxExt<'tcx> {
    // Provided methods
    fn read_from_host(
        &mut self,
        file: impl Read,
        len: usize,
        ptr: Pointer,
    ) -> InterpResult<'tcx, Result<usize, IoError>> { ... }
    fn write_to_host(
        &mut self,
        file: impl Write,
        len: usize,
        ptr: Pointer,
    ) -> InterpResult<'tcx, Result<usize, IoError>> { ... }
}

Provided Methods

fn read_from_host( &mut self, file: impl Read, len: usize, ptr: Pointer, ) -> InterpResult<'tcx, Result<usize, IoError>>

Read data from a host Read type, store the result into machine memory, and return whether that worked.

fn write_to_host( &mut self, file: impl Write, len: usize, ptr: Pointer, ) -> InterpResult<'tcx, Result<usize, IoError>>

Write data to a host Write type, withthe bytes taken from machine memory.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.