miri::shims
Module native_lib

Implements calling functions from a native library.

Enums

CArg 🔒

Enum of supported arguments to external C functions.

Traits

EvalContextExt

EvalContextExtPriv 🔒

Functions

imm_to_carg 🔒

Extract the scalar value from the result of reading a scalar from the machine, and convert it to a CArg.