miri::shims::native_lib::trace::child

Struct Supervisor

pub struct Supervisor {
    message_tx: IpcSender<TraceRequest>,
    confirm_rx: IpcReceiver<Confirmation>,
    event_rx: IpcReceiver<MemEvents>,
}

The main means of communication between the child and parent process, allowing the former to send requests and get info from the latter.

Fields
message_tx: IpcSender<TraceRequest>

Sender for FFI-mode-related requests.

confirm_rx: IpcReceiver<Confirmation>

Used for synchronisation, allowing us to receive confirmation that the parent process has handled the request from message_tx.

event_rx: IpcReceiver<MemEvents>

Receiver for memory acceses that ocurred during the FFI call.

Implementations

impl Supervisor

pub fn is_enabled() -> bool

Returns true if the supervisor process exists, and false otherwise.

unsafe fn protect_pages( pages: impl Iterator<Item = (NonNull<u8>, usize)>, prot: ProtFlags, ) -> Result<(), Errno>

pub fn do_ffi<'tcx>( alloc: &Rc<RefCell<IsolatedAlloc>>, f: impl FnOnce() -> InterpResult<'tcx, ImmTy<'tcx>>, ) -> InterpResult<'tcx, (ImmTy<'tcx>, Option<MemEvents>)>

Performs an arbitrary FFI call, enabling tracing from the supervisor. As this locks the supervisor via a mutex, no other threads may enter FFI until this function returns.