miri::shims::native_lib::trace::parent

Enum ExecEvent

pub enum ExecEvent {
    Start(StartFfiInfo),
    End,
    Status(Pid, Signal),
    Syscall(Pid),
    Died(Option<i32>),
}

A unified event representing something happening on the child process. Wraps nix’s WaitStatus and our custom signals so it can all be done with one match statement.

Variants

Start(StartFfiInfo)

Child process requests that we begin monitoring it.

End

Child requests that we stop monitoring and pass over the events we detected.

Status(Pid, Signal)

The child process with the specified pid was stopped by the given signal.

Syscall(Pid)

The child process with the specified pid entered or existed a syscall.

Died(Option<i32>)

A child process exited or was killed; if we have a return code, it is specified.