miri::shims::native_lib::trace::parent

Function capstone_find_events

fn capstone_find_events(
    addr: usize,
    op: &ArchOperand,
    acc_events: &mut Vec<AccessEvent>,
) -> bool

Add the memory events from op being executed while there is a memory access at addr to acc_events. Return whether this was a memory operand.