miri::shims::native_lib::trace::parent

Function handle_segfault

fn handle_segfault(
    pid: Pid,
    ch_pages: &[usize],
    ch_stack: usize,
    page_size: usize,
    cs: &Capstone,
    acc_events: &mut Vec<AccessEvent>,
) -> Result<(), ExecEnd>

Grabs the access that caused a segfault and logs it down if it’s to our memory, or kills the child and returns the appropriate error otherwise.