miri::shims::native_lib::trace::parent

Function wait_for_signal

fn wait_for_signal(
    pid: Option<Pid>,
    wait_signal: Signal,
    init_cont: InitialCont,
) -> Result<Pid, ExecEnd>

Waits for wait_signal. If init_cont, it will first do a ptrace::cont. We want to avoid that in some cases, like at the beginning of FFI.

If pid is None, only one wait will be done and init_cont should be false.