miri::shims::native_lib::trace::parent

Trait ArchIndependentRegs

trait ArchIndependentRegs {
    // Required methods
    fn ip(&self) -> usize;
    fn set_ip(&mut self, ip: usize);
    fn set_sp(&mut self, sp: usize);
}

Allows us to get common arguments from the user_regs_t across architectures. Normally this would land us ABI hell, but thankfully all of our usecases consist of functions with a small number of register-sized integer arguments. See https://man7.org/linux/man-pages/man2/syscall.2.html for sources.

Required Methods

fn ip(&self) -> usize

Gets the address of the instruction pointer.

fn set_ip(&mut self, ip: usize)

Set the instruction pointer; remember to also set the stack pointer, or else the stack might get messed up!

fn set_sp(&mut self, sp: usize)

Set the stack pointer, ideally to a zeroed-out area.