miri::shims::native_lib

Trait EvalContextExtPriv

trait EvalContextExtPriv<'tcx>: MiriInterpCxExt<'tcx> {
    // Provided methods
    fn call_native_with_args<'a>(
        &mut self,
        link_name: Symbol,
        dest: &MPlaceTy<'tcx>,
        ptr: CodePtr,
        libffi_args: Vec<Arg<'a>>,
    ) -> InterpResult<'tcx, (ImmTy<'tcx>, Option<MemEvents>)> { ... }
    fn get_func_ptr_explicitly_from_lib(
        &mut self,
        link_name: Symbol,
    ) -> Option<CodePtr> { ... }
    fn tracing_apply_accesses(
        &mut self,
        events: MemEvents,
    ) -> InterpResult<'tcx> { ... }
}

Provided Methods

fn call_native_with_args<'a>( &mut self, link_name: Symbol, dest: &MPlaceTy<'tcx>, ptr: CodePtr, libffi_args: Vec<Arg<'a>>, ) -> InterpResult<'tcx, (ImmTy<'tcx>, Option<MemEvents>)>

Call native host function and return the output as an immediate.

fn get_func_ptr_explicitly_from_lib( &mut self, link_name: Symbol, ) -> Option<CodePtr>

Get the pointer to the function of the specified name in the shared object file, if it exists. The function must be in one of the shared object files specified: we do not return pointers to functions in dependencies of libraries.

fn tracing_apply_accesses(&mut self, events: MemEvents) -> InterpResult<'tcx>

Applies the events to Miri’s internal state. The event vector must be ordered sequentially by when the accesses happened, and the sizes are assumed to be exact.