Struct miri::shims::tls::TlsData

pub struct TlsData<'tcx> {
    next_key: u128,
    keys: BTreeMap<u128, TlsEntry<'tcx>>,
    macos_thread_dtors: BTreeMap<ThreadId, (Instance<'tcx>, Scalar<Provenance>)>,
}

Fields
next_key: u128

The Key to use for the next thread-local allocation.

keys: BTreeMap<u128, TlsEntry<'tcx>>

pthreads-style thread-local storage.

macos_thread_dtors: BTreeMap<ThreadId, (Instance<'tcx>, Scalar<Provenance>)>

A single per thread destructor of the thread local storage (that’s how things work on macOS) with a data argument.

Implementations

impl<'tcx> TlsData<'tcx>

pub fn create_tls_key( &mut self, dtor: Option<Instance<'tcx>>, max_size: Size ) -> InterpResult<'tcx, u128>

Generate a new TLS key with the given destructor. max_size determines the integer size the key has to fit in.

pub fn delete_tls_key(&mut self, key: u128) -> InterpResult<'tcx>

pub fn load_tls( &self, key: u128, thread_id: ThreadId, cx: &impl HasDataLayout ) -> InterpResult<'tcx, Scalar<Provenance>>

pub fn store_tls( &mut self, key: u128, thread_id: ThreadId, new_data: Scalar<Provenance>, cx: &impl HasDataLayout ) -> InterpResult<'tcx>

pub fn set_macos_thread_dtor( &mut self, thread: ThreadId, dtor: Instance<'tcx>, data: Scalar<Provenance> ) -> InterpResult<'tcx>

Set the thread wide destructor of the thread local storage for the given thread. This function is used to implement _tlv_atexit shim on MacOS.

Thread wide dtors are available only on MacOS. There is one destructor per thread as can be guessed from the following comment in the _tlv_atexit implementation:

NOTE: this does not need locks because it only operates on current thread data

fn fetch_tls_dtor( &mut self, key: Option<u128>, thread_id: ThreadId ) -> Option<(Instance<'tcx>, Scalar<Provenance>, u128)>

Returns a dtor, its argument and its index, if one is supposed to run. key is the last dtors that was run; we return the next one after that.

An optional destructor function may be associated with each key value. At thread exit, if a key value has a non-NULL destructor pointer, and the thread has a non-NULL value associated with that key, the value of the key is set to NULL, and then the function pointed to is called with the previously associated value as its sole argument. The order of destructor calls is unspecified if more than one destructor exists for a thread when it exits.

If, after all the destructors have been called for all non-NULL values with associated destructors, there are still some non-NULL values with associated destructors, then the process is repeated. If, after at least {PTHREAD_DESTRUCTOR_ITERATIONS} iterations of destructor calls for outstanding non-NULL values, there are still some non-NULL values with associated destructors, implementations may stop calling destructors, or they may continue calling destructors until no non-NULL values with associated destructors exist, even though this might result in an infinite loop.

fn delete_all_thread_tls(&mut self, thread_id: ThreadId)

Delete all TLS entries for the given thread. This function should be called after all TLS destructors have already finished.