Trait miri::shims::EvalContextExt

pub trait EvalContextExt<'mir, 'tcx: 'mir>: MiriEvalContextExt<'mir, 'tcx> {
    fn find_mir_or_eval_fn(
        &mut self,
        instance: Instance<'tcx>,
        abi: Abi,
        args: &[OpTy<'tcx, Provenance>],
        dest: &PlaceTy<'tcx, Provenance>,
        ret: Option<BasicBlock>,
        unwind: StackPopUnwind
    ) -> InterpResult<'tcx, Option<(&'mir Body<'tcx>, Instance<'tcx>)>> { ... }
    fn align_offset(
        &mut self,
        ptr_op: &OpTy<'tcx, Provenance>,
        align_op: &OpTy<'tcx, Provenance>,
        dest: &PlaceTy<'tcx, Provenance>,
        ret: Option<BasicBlock>,
        unwind: StackPopUnwind
    ) -> InterpResult<'tcx, bool> { ... }
}

Provided Methods

fn find_mir_or_eval_fn(
&mut self,
instance: Instance<'tcx>,
abi: Abi,
args: &[OpTy<'tcx, Provenance>],
dest: &PlaceTy<'tcx, Provenance>,
ret: Option<BasicBlock>,
unwind: StackPopUnwind
) -> InterpResult<'tcx, Option<(&'mir Body<'tcx>, Instance<'tcx>)>>

fn align_offset(
&mut self,
ptr_op: &OpTy<'tcx, Provenance>,
align_op: &OpTy<'tcx, Provenance>,
dest: &PlaceTy<'tcx, Provenance>,
ret: Option<BasicBlock>,
unwind: StackPopUnwind
) -> InterpResult<'tcx, bool>

Returns true if the computation was performed, and false if we should just evaluate the actual MIR of align_offset.