Module miri::shims::unix

Re-exports

pub use env::UnixEnvVars;

pub use fd::FdTable;

pub use fd::FileDescription;

pub use fs::DirTable;

pub use env::EvalContextExt as _;

pub use fd::EvalContextExt as _;

pub use fs::EvalContextExt as _;

pub use mem::EvalContextExt as _;

pub use socket::EvalContextExt as _;

pub use sync::EvalContextExt as _;

pub use thread::EvalContextExt as _;

Modules

android 🔒

env 🔒

fd 🔒

General management of file descriptors, and support for standard file descriptors (stdin/stdout/stderr).

foreign_items

freebsd 🔒

fs 🔒

File and file system access

linux 🔒

macos 🔒

mem 🔒

This is an incomplete implementation of mmap/munmap which is restricted in order to be implementable on top of the existing memory system. The point of these function as-written is to allow memory allocators written entirely in Rust to be executed by Miri. This implementation does not support other uses of mmap such as file mappings.

socket 🔒

solarish 🔒

sync 🔒

thread 🔒

Constants

UID 🔒