Trait miri::shims::unix::fs::EvalContextExtPrivate

trait EvalContextExtPrivate<'mir, 'tcx: 'mir>: MiriInterpCxExt<'mir, 'tcx> {
    fn macos_stat_write_buf(
        &mut self,
        metadata: FileMetadata,
        buf_op: &OpTy<'tcx, Provenance>
    ) -> InterpResult<'tcx, i32> { ... }
    fn handle_not_found<T: From<i32>>(&mut self) -> InterpResult<'tcx, T> { ... }
    fn file_type_to_d_type(
        &mut self,
        file_type: Result<FileType>
    ) -> InterpResult<'tcx, i32> { ... }
}

Provided Methods

fn macos_stat_write_buf(
&mut self,
metadata: FileMetadata,
buf_op: &OpTy<'tcx, Provenance>
) -> InterpResult<'tcx, i32>

fn handle_not_found<T: From<i32>>(&mut self) -> InterpResult<'tcx, T>

Function used when a handle is not found inside FileHandler. It returns Ok(-1)and sets the last OS error to libc::EBADF (invalid file descriptor). This function uses T: From<i32> instead of i32 directly because some fs functions return different integer types (like read, that returns an i64).

fn file_type_to_d_type(
&mut self,
file_type: Result<FileType>
) -> InterpResult<'tcx, i32>