miri::shims::unix::linux
Module epoll

Structs

Epoll 🔒

An Epoll file descriptor connects file handles and epoll events

EpollEventInstance

EpollEventInstance contains information that will be returned by epoll_wait.

EpollEventInterest

EpollEventInterest registers the file description information to an epoll instance during a successful epoll_ctl call. It also stores additional information needed to check and update readiness state for epoll_wait.

EpollInterestTable

The table of all EpollEventInterest. The BTreeMap key is the FdId of an active file description registered with any epoll instance. The value is a list of EpollEventInterest associated with that file description.

EpollReadyEvents

EpollReadyEvents reflects the readiness of a file description.

Traits

EvalContextExt

Functions

blocking_epoll_callback 🔒

Callback function after epoll_wait unblocks

check_and_update_one_event_interest 🔒

This helper function checks whether an epoll notification should be triggered for a specific epoll_interest and, if necessary, triggers the notification, and returns whether the notification was added/updated. Unlike check_and_update_readiness, this function sends a notification to only one epoll instance.

ready_list_next 🔒

This function takes in ready list and returns EpollEventInstance with file description that is not closed.