Struct miri::shims::unix::linux::eventfd::Event

struct Event {
    counter: u64,
    is_nonblock: bool,
    clock: VClock,
}

A kind of file descriptor created by eventfd. The Event type isn’t currently written to by eventfd. The interface is meant to keep track of objects associated with a file descriptor. For more information see the man page below:

https://man.netbsd.org/eventfd.2

Fields
counter: u64

The object contains an unsigned 64-bit integer (uint64_t) counter that is maintained by the kernel. This counter is initialized with the value specified in the argument initval.

is_nonblock: boolclock: VClock