miri::shims::x86

Function test_high_bits_masked

fn test_high_bits_masked<'tcx>(
    ecx: &MiriInterpCx<'tcx>,
    op: &OpTy<'tcx>,
    mask: &OpTy<'tcx>,
) -> InterpResult<'tcx, (bool, bool)>

Calculates two booleans.

The first is true when the highest bit of each element of op & mask is zero. The second is true when the highest bit of each element of !op & mask is zero.