miri::shims::x86
Module gfni

Statics

TABLE 🔒

A lookup table for computing the inverse byte for the inverse affine transformation. See https://www.corsix.org/content/galois-field-instructions-2021-cpus for the definition of gf_inv which was used for the creation of this table.

Traits

EvalContextExt 🔒

Functions

affine_transform 🔒

Calculates the affine transformation right * left + imm8 inside the finite field GF(2^8). right is an 8x8 bit matrix, left and imm8 are bit vectors. If inverse is set, then the inverse transformation with respect to the reduction polynomial x^8 + x^4 + x^3 + x + 1 is performed instead.

gf2p8_mul 🔒

Multiplies packed 8-bit integers in left and right in the finite field GF(2^8) and store the results in dst. The field GF(2^8) is represented in polynomial representation with the reduction polynomial x^8 + x^4 + x^3 + x + 1. See https://www.corsix.org/content/galois-field-instructions-2021-cpus for details.