Module miri::shims::x86::sse

Enums

FloatBinOp 🔒

FloatCmpOp 🔒

Floating point comparison operation

FloatUnaryOp 🔒

Traits

EvalContextExt

Functions

apply_random_float_error 🔒

Disturbes a floating-point result by a relative error on the order of (-2^scale, 2^scale).

bin_op_f32 🔒

Performs which scalar operation on left and right and returns the result.

bin_op_ps 🔒

Performs which operation on each component of left, and right storing the result is stored in dest.

bin_op_ss 🔒

Performs which operation on the first component of left and right and copies the other components from left. The result is stored in dest.

unary_op_f32 🔒

Performs which scalar operation on op and returns the result.

unary_op_ps 🔒

Performs which operation on each component of op, storing the result is stored in dest.

unary_op_ss 🔒

Performs which operation on the first component of op and copies the other components. The result is stored in dest.