Struct miri::stacked_borrows::Stacks

pub struct Stacks {
    stacks: RangeMap<Stack>,
    history: AllocHistory,
    exposed_tags: FxHashSet<SbTag>,
}

Extra per-allocation state.

Fields
stacks: RangeMap<Stack>history: AllocHistory

Stores past operations on this allocation

exposed_tags: FxHashSet<SbTag>

The set of tags that have been exposed inside this allocation.

Implementations

impl<'tcx> Stacks

Map per-stack operations to higher-level per-location-range operations.

fn new(
size: Size,
perm: Permission,
tag: SbTag,
id: AllocId,
current_span: &mut CurrentSpan<'_, '_, '_>
) -> Self

Creates a new stack with an initial tag. For diagnostic purposes, we also need to know the AllocId of the allocation this is associated with.

fn for_each(
&mut self,
range: AllocRange,
dcx_builder: DiagnosticCxBuilder<'_, '_, '_, 'tcx>,
f: impl FnMut(&mut Stack, &mut DiagnosticCx<'_, '_, '_, '_, 'tcx>, &mut FxHashSet<SbTag>) -> InterpResult<'tcx>
) -> InterpResult<'tcx>

Call f on every stack in the range.

impl Stacks

Glue code to connect with Miri Machine Hooks

pub fn new_allocation(
id: AllocId,
size: Size,
state: &RefCell<GlobalStateInner>,
kind: MemoryKind<MiriMemoryKind>,
current_span: CurrentSpan<'_, '_, '_>
) -> Self

pub fn before_memory_read<'tcx, 'mir, 'ecx>(
&mut self,
alloc_id: AllocId,
tag: ProvenanceExtra,
range: AllocRange,
state: &RefCell<GlobalStateInner>,
current_span: CurrentSpan<'ecx, 'mir, 'tcx>,
threads: &'ecx ThreadManager<'mir, 'tcx>
) -> InterpResult<'tcx>where
'tcx: 'ecx,

pub fn before_memory_write<'tcx, 'mir, 'ecx>(
&mut self,
alloc_id: AllocId,
tag: ProvenanceExtra,
range: AllocRange,
state: &RefCell<GlobalStateInner>,
current_span: CurrentSpan<'ecx, 'mir, 'tcx>,
threads: &'ecx ThreadManager<'mir, 'tcx>
) -> InterpResult<'tcx>

pub fn before_memory_deallocation<'tcx, 'mir, 'ecx>(
&mut self,
alloc_id: AllocId,
tag: ProvenanceExtra,
range: AllocRange,
state: &RefCell<GlobalStateInner>,
current_span: CurrentSpan<'ecx, 'mir, 'tcx>,
threads: &'ecx ThreadManager<'mir, 'tcx>
) -> InterpResult<'tcx>