Struct miri::MiriConfig

pub struct MiriConfig {

    pub env: Vec<(OsString, OsString)>,
    pub validate: bool,
    pub stacked_borrows: bool,
    pub check_alignment: AlignmentCheck,
    pub check_abi: bool,
    pub isolated_op: IsolatedOp,
    pub ignore_leaks: bool,
    pub excluded_env_vars: Vec<String>,
    pub forwarded_env_vars: Vec<String>,
    pub args: Vec<String>,
    pub seed: Option<u64>,
    pub tracked_pointer_tags: FxHashSet<SbTag>,
    pub tracked_call_ids: FxHashSet<CallId>,
    pub tracked_alloc_ids: FxHashSet<AllocId>,
    pub data_race_detector: bool,
    pub weak_memory_emulation: bool,
    pub track_outdated_loads: bool,
    pub cmpxchg_weak_failure_rate: f64,
    pub measureme_out: Option<String>,
    pub panic_on_unsupported: bool,
    pub backtrace_style: BacktraceStyle,
    pub provenance_mode: ProvenanceMode,
    pub mute_stdout_stderr: bool,
    pub preemption_rate: f64,
    pub report_progress: Option<u32>,
    pub retag_fields: bool,
    pub external_so_file: Option<PathBuf>,

}

Configuration needed to spawn a Miri instance.

Fields
env: Vec<(OsString, OsString)>

The host environment snapshot to use as basis for what is provided to the interpreted program. (This is still subject to isolation as well as excluded_env_vars and forwarded_env_vars.)

validate: bool

Determine if validity checking is enabled.

stacked_borrows: bool

Determines if Stacked Borrows is enabled.

check_alignment: AlignmentCheck

Controls alignment checking.

check_abi: bool

Controls function ABI checking.

isolated_op: IsolatedOp

Action for an op requiring communication with the host.

ignore_leaks: bool

Determines if memory leaks should be ignored.

excluded_env_vars: Vec<String>

Environment variables that should always be isolated from the host.

forwarded_env_vars: Vec<String>

Environment variables that should always be forwarded from the host.

args: Vec<String>

Command-line arguments passed to the interpreted program.

seed: Option<u64>

The seed to use when non-determinism or randomness are required (e.g. ptr-to-int cast, getrandom()).

tracked_pointer_tags: FxHashSet<SbTag>

The stacked borrows pointer ids to report about

tracked_call_ids: FxHashSet<CallId>

The stacked borrows call IDs to report about

tracked_alloc_ids: FxHashSet<AllocId>

The allocation ids to report about.

data_race_detector: bool

Determine if data race detection should be enabled

weak_memory_emulation: bool

Determine if weak memory emulation should be enabled. Requires data race detection to be enabled

track_outdated_loads: bool

Track when an outdated (weak memory) load happens.

cmpxchg_weak_failure_rate: f64

Rate of spurious failures for compare_exchange_weak atomic operations, between 0.0 and 1.0, defaulting to 0.8 (80% chance of failure).

measureme_out: Option<String>

If Some, enable the measureme profiler, writing results to a file with the specified prefix.

panic_on_unsupported: bool

Panic when unsupported functionality is encountered.

backtrace_style: BacktraceStyle

Which style to use for printing backtraces.

provenance_mode: ProvenanceMode

Which provenance to use for int2ptr casts

mute_stdout_stderr: bool

Whether to ignore any output by the program. This is helpful when debugging miri as its messages don’t get intermingled with the program messages.

preemption_rate: f64

The probability of the active thread being preempted at the end of each basic block.

report_progress: Option<u32>

Report the current instruction being executed every N basic blocks.

retag_fields: bool

Whether Stacked Borrows retagging should recurse into fields of datatypes.

external_so_file: Option<PathBuf>

The location of a shared object file to load when calling external functions FIXME! consider allowing users to specify paths to multiple SO files, or to a directory