Struct miri::Tree

pub struct Tree { /* private fields */ }

Tree structure with both parents and children since we want to be able to traverse the tree efficiently in both directions.

Implementations

impl<'tcx> Tree

pub fn give_pointer_debug_name( &mut self, tag: BorTag, nth_parent: u8, name: &str ) -> InterpResult<'tcx>

Debug helper: assign name to tag.

pub fn is_allocation_of(&self, tag: BorTag) -> bool

Debug helper: determines if the tree contains a tag.

impl<'tcx> Tree

pub fn print_tree( &self, protected_tags: &FxHashMap<BorTag, ProtectorKind>, show_unnamed: bool ) -> InterpResult<'tcx>

Display the contents of the tree.

impl Tree

pub fn new(root_tag: BorTag, size: Size) -> Self

Create a new tree, with only a root pointer.

impl<'tcx> Tree

pub fn new_child( &mut self, parent_tag: BorTag, new_tag: BorTag, default_initial_perm: Permission, range: AllocRange ) -> InterpResult<'tcx>

Insert a new tag in the tree

pub fn dealloc( &mut self, tag: BorTag, range: AllocRange, global: &RefCell<GlobalStateInner> ) -> InterpResult<'tcx>

Deallocation requires

a pointer that permits write accesses

the absence of Strong Protectors anywhere in the allocation

pub fn perform_access( &mut self, access_kind: AccessKind, tag: BorTag, range: AllocRange, global: &RefCell<GlobalStateInner> ) -> InterpResult<'tcx>

Maps the following propagation procedure to each range:

initialize if needed;

compute new state after transition;

check that there is no protector that would forbid this;

record this specific location as accessed.

impl Tree

Integration with the BorTag garbage collector

pub fn remove_unreachable_tags(&mut self, live_tags: &FxHashSet<BorTag>)

impl<'tcx> Tree

pub fn new_allocation( id: AllocId, size: Size, state: &mut GlobalStateInner, _kind: MemoryKind<MiriMemoryKind>, machine: &MiriMachine<'_, 'tcx> ) -> Self

Create a new allocation, i.e. a new tree

pub fn before_memory_access( &mut self, access_kind: AccessKind, alloc_id: AllocId, prov: ProvenanceExtra, range: AllocRange, machine: &MiriMachine<'_, 'tcx> ) -> InterpResult<'tcx>

Check that an access on the entire range is permitted, and update the tree.

pub fn before_memory_deallocation( &mut self, _alloc_id: AllocId, prov: ProvenanceExtra, range: AllocRange, machine: &MiriMachine<'_, 'tcx> ) -> InterpResult<'tcx>

Check that this pointer has permission to deallocate this range.

pub fn expose_tag(&mut self, _tag: BorTag)