Trait miri::Machine

pub trait Machine<'mir, 'tcx> {
    type MemoryKind: 'static + Debug + Display + MayLeak + Eq;
    type Provenance: 'static + Provenance + Eq + Hash;
    type ProvenanceExtra: 'static + Copy;
    type ExtraFnVal: Debug + Copy;
    type FrameExtra;
    type AllocExtra: 'static + Debug + Clone;
    type MemoryMap: AllocMap<AllocId, (MemoryKind<Self::MemoryKind>, Allocation<Self::Provenance, Self::AllocExtra>)> + Default + Clone;

    const GLOBAL_KIND: Option<Self::MemoryKind>;
    const PANIC_ON_ALLOC_FAIL: bool;

    fn enforce_alignment(ecx: &InterpCx<'mir, 'tcx, Self>) -> bool;
    fn use_addr_for_alignment_check(ecx: &InterpCx<'mir, 'tcx, Self>) -> bool;
    fn enforce_validity(ecx: &InterpCx<'mir, 'tcx, Self>) -> bool;
    fn checked_binop_checks_overflow(_ecx: &InterpCx<'mir, 'tcx, Self>) -> bool;
    fn find_mir_or_eval_fn(
        ecx: &mut InterpCx<'mir, 'tcx, Self>,
        instance: Instance<'tcx>,
        abi: Abi,
        args: &[OpTy<'tcx, Self::Provenance>],
        destination: &PlaceTy<'tcx, Self::Provenance>,
        target: Option<BasicBlock>,
        unwind: StackPopUnwind
    ) -> Result<Option<(&'mir Body<'tcx>, Instance<'tcx>)>, InterpErrorInfo<'tcx>>;
    fn call_extra_fn(
        ecx: &mut InterpCx<'mir, 'tcx, Self>,
        fn_val: Self::ExtraFnVal,
        abi: Abi,
        args: &[OpTy<'tcx, Self::Provenance>],
        destination: &PlaceTy<'tcx, Self::Provenance>,
        target: Option<BasicBlock>,
        unwind: StackPopUnwind
    ) -> Result<(), InterpErrorInfo<'tcx>>;
    fn call_intrinsic(
        ecx: &mut InterpCx<'mir, 'tcx, Self>,
        instance: Instance<'tcx>,
        args: &[OpTy<'tcx, Self::Provenance>],
        destination: &PlaceTy<'tcx, Self::Provenance>,
        target: Option<BasicBlock>,
        unwind: StackPopUnwind
    ) -> Result<(), InterpErrorInfo<'tcx>>;
    fn assert_panic(
        ecx: &mut InterpCx<'mir, 'tcx, Self>,
        msg: &AssertKind<Operand<'tcx>>,
        unwind: Option<BasicBlock>
    ) -> Result<(), InterpErrorInfo<'tcx>>;
    fn binary_ptr_op(
        ecx: &InterpCx<'mir, 'tcx, Self>,
        bin_op: BinOp,
        left: &ImmTy<'tcx, Self::Provenance>,
        right: &ImmTy<'tcx, Self::Provenance>
    ) -> Result<(Scalar<Self::Provenance>, bool, Ty<'tcx>), InterpErrorInfo<'tcx>>;
    fn extern_static_base_pointer(
        ecx: &InterpCx<'mir, 'tcx, Self>,
        def_id: DefId
    ) -> Result<Pointer<Self::Provenance>, InterpErrorInfo<'tcx>>;
    fn adjust_alloc_base_pointer(
        ecx: &InterpCx<'mir, 'tcx, Self>,
        ptr: Pointer<AllocId>
    ) -> Pointer<Self::Provenance>;
    fn ptr_from_addr_cast(
        ecx: &InterpCx<'mir, 'tcx, Self>,
        addr: u64
    ) -> Result<Pointer<Option<Self::Provenance>>, InterpErrorInfo<'tcx>>;
    fn expose_ptr(
        ecx: &mut InterpCx<'mir, 'tcx, Self>,
        ptr: Pointer<Self::Provenance>
    ) -> Result<(), InterpErrorInfo<'tcx>>;
    fn ptr_get_alloc(
        ecx: &InterpCx<'mir, 'tcx, Self>,
        ptr: Pointer<Self::Provenance>
    ) -> Option<(AllocId, Size, Self::ProvenanceExtra)>;
    fn adjust_allocation(
        ecx: &InterpCx<'mir, 'tcx, Self>,
        id: AllocId,
        alloc: Cow<'b, Allocation<AllocId, ()>>,
        kind: Option<MemoryKind<Self::MemoryKind>>
    ) -> Result<Cow<'b, Allocation<Self::Provenance, Self::AllocExtra>>, InterpErrorInfo<'tcx>>;
    fn init_frame_extra(
        ecx: &mut InterpCx<'mir, 'tcx, Self>,
        frame: Frame<'mir, 'tcx, Self::Provenance, ()>
    ) -> Result<Frame<'mir, 'tcx, Self::Provenance, Self::FrameExtra>, InterpErrorInfo<'tcx>>;
    fn stack(
        ecx: &'a InterpCx<'mir, 'tcx, Self>
    ) -> &'a [Frame<'mir, 'tcx, Self::Provenance, Self::FrameExtra>];
    fn stack_mut(
        ecx: &'a mut InterpCx<'mir, 'tcx, Self>
    ) -> &'a mut Vec<Frame<'mir, 'tcx, Self::Provenance, Self::FrameExtra>, Global>;

    fn enforce_abi(_ecx: &InterpCx<'mir, 'tcx, Self>) -> bool { ... }
    fn load_mir(
        ecx: &InterpCx<'mir, 'tcx, Self>,
        instance: InstanceDef<'tcx>
    ) -> Result<&'tcx Body<'tcx>, InterpErrorInfo<'tcx>> { ... }
    fn abort(
        _ecx: &mut InterpCx<'mir, 'tcx, Self>,
        _msg: String
    ) -> Result<!, InterpErrorInfo<'tcx>> { ... }
    fn access_local_mut(
        ecx: &'a mut InterpCx<'mir, 'tcx, Self>,
        frame: usize,
        local: Local
    ) -> Result<&'a mut Operand<Self::Provenance>, InterpErrorInfo<'tcx>>
   where
        'tcx: 'mir,
    { ... }
    fn before_terminator(
        _ecx: &mut InterpCx<'mir, 'tcx, Self>
    ) -> Result<(), InterpErrorInfo<'tcx>> { ... }
    fn before_access_global(
        _tcx: TyCtxt<'tcx>,
        _machine: &Self,
        _alloc_id: AllocId,
        _allocation: ConstAllocation<'tcx, AllocId, ()>,
        _static_def_id: Option<DefId>,
        _is_write: bool
    ) -> Result<(), InterpErrorInfo<'tcx>> { ... }
    fn thread_local_static_base_pointer(
        _ecx: &mut InterpCx<'mir, 'tcx, Self>,
        def_id: DefId
    ) -> Result<Pointer<Self::Provenance>, InterpErrorInfo<'tcx>> { ... }
    fn eval_inline_asm(
        _ecx: &mut InterpCx<'mir, 'tcx, Self>,
        _template: &'tcx [InlineAsmTemplatePiece],
        _operands: &[InlineAsmOperand<'tcx>],
        _options: InlineAsmOptions
    ) -> Result<(), InterpErrorInfo<'tcx>> { ... }
    fn before_memory_read(
        _tcx: TyCtxt<'tcx>,
        _machine: &Self,
        _alloc_extra: &Self::AllocExtra,
        _prov: (AllocId, Self::ProvenanceExtra),
        _range: AllocRange
    ) -> Result<(), InterpErrorInfo<'tcx>> { ... }
    fn before_memory_write(
        _tcx: TyCtxt<'tcx>,
        _machine: &mut Self,
        _alloc_extra: &mut Self::AllocExtra,
        _prov: (AllocId, Self::ProvenanceExtra),
        _range: AllocRange
    ) -> Result<(), InterpErrorInfo<'tcx>> { ... }
    fn before_memory_deallocation(
        _tcx: TyCtxt<'tcx>,
        _machine: &mut Self,
        _alloc_extra: &mut Self::AllocExtra,
        _prov: (AllocId, Self::ProvenanceExtra),
        _range: AllocRange
    ) -> Result<(), InterpErrorInfo<'tcx>> { ... }
    fn retag(
        _ecx: &mut InterpCx<'mir, 'tcx, Self>,
        _kind: RetagKind,
        _place: &PlaceTy<'tcx, Self::Provenance>
    ) -> Result<(), InterpErrorInfo<'tcx>> { ... }
    fn after_stack_push(
        _ecx: &mut InterpCx<'mir, 'tcx, Self>
    ) -> Result<(), InterpErrorInfo<'tcx>> { ... }
    fn after_stack_pop(
        _ecx: &mut InterpCx<'mir, 'tcx, Self>,
        _frame: Frame<'mir, 'tcx, Self::Provenance, Self::FrameExtra>,
        unwinding: bool
    ) -> Result<StackPopJump, InterpErrorInfo<'tcx>> { ... }

}

Methods of this trait signifies a point where CTFE evaluation would fail and some use case dependent behaviour can instead be applied.

Required Associated Types

type MemoryKind: 'static + Debug + Display + MayLeak + Eq

Additional memory kinds a machine wishes to distinguish from the builtin ones

type Provenance: 'static + Provenance + Eq + Hash

Pointers are “tagged” with provenance information; typically the AllocId they belong to.

type ProvenanceExtra: 'static + Copy

When getting the AllocId of a pointer, some extra data is also obtained from the provenance that is passed to memory access hooks so they can do things with it.

type ExtraFnVal: Debug + Copy

Machines can define extra (non-instance) things that represent values of function pointers. For example, Miri uses this to return a function pointer from dlsym that can later be called to execute the right thing.

type FrameExtra

Extra data stored in every call frame.

type AllocExtra: 'static + Debug + Clone

Extra data stored in every allocation.

type MemoryMap: AllocMap<AllocId, (MemoryKind<Self::MemoryKind>, Allocation<Self::Provenance, Self::AllocExtra>)> + Default + Clone

Memory’s allocation map

Required Associated Constants

const GLOBAL_KIND: Option<Self::MemoryKind>

The memory kind to use for copied global memory (held in tcx) – or None if such memory should not be mutated and thus any such attempt will cause a ModifiedStatic error to be raised. Statics are copied under two circumstances: When they are mutated, and when adjust_allocation (see below) returns an owned allocation that is added to the memory so that the work is not done twice.

const PANIC_ON_ALLOC_FAIL: bool

Should the machine panic on allocation failures?

Required Methods

fn enforce_alignment(ecx: &InterpCx<'mir, 'tcx, Self>) -> bool

Whether memory accesses should be alignment-checked.

fn use_addr_for_alignment_check(ecx: &InterpCx<'mir, 'tcx, Self>) -> bool

Whether, when checking alignment, we should look at the actual address and thus support custom alignment logic based on whatever the integer address happens to be.

If this returns true, Provenance::OFFSET_IS_ADDR must be true.

fn enforce_validity(ecx: &InterpCx<'mir, 'tcx, Self>) -> bool

Whether to enforce the validity invariant

fn checked_binop_checks_overflow(_ecx: &InterpCx<'mir, 'tcx, Self>) -> bool

Whether CheckedBinOp MIR statements should actually check for overflow.

fn find_mir_or_eval_fn(
ecx: &mut InterpCx<'mir, 'tcx, Self>,
instance: Instance<'tcx>,
abi: Abi,
args: &[OpTy<'tcx, Self::Provenance>],
destination: &PlaceTy<'tcx, Self::Provenance>,
target: Option<BasicBlock>,
unwind: StackPopUnwind
) -> Result<Option<(&'mir Body<'tcx>, Instance<'tcx>)>, InterpErrorInfo<'tcx>>

Entry point to all function calls.

Returns either the mir to use for the call, or None if execution should just proceed (which usually means this hook did all the work that the called function should usually have done). In the latter case, it is this hook’s responsibility to advance the instruction pointer! (This is to support functions like __rust_maybe_catch_panic that neither find a MIR nor just jump to ret, but instead push their own stack frame.) Passing destand ret in the same Option proved very annoying when only one of them was used.

fn call_extra_fn(
ecx: &mut InterpCx<'mir, 'tcx, Self>,
fn_val: Self::ExtraFnVal,
abi: Abi,
args: &[OpTy<'tcx, Self::Provenance>],
destination: &PlaceTy<'tcx, Self::Provenance>,
target: Option<BasicBlock>,
unwind: StackPopUnwind
) -> Result<(), InterpErrorInfo<'tcx>>

Execute fn_val. It is the hook’s responsibility to advance the instruction pointer as appropriate.

fn call_intrinsic(
ecx: &mut InterpCx<'mir, 'tcx, Self>,
instance: Instance<'tcx>,
args: &[OpTy<'tcx, Self::Provenance>],
destination: &PlaceTy<'tcx, Self::Provenance>,
target: Option<BasicBlock>,
unwind: StackPopUnwind
) -> Result<(), InterpErrorInfo<'tcx>>

Directly process an intrinsic without pushing a stack frame. It is the hook’s responsibility to advance the instruction pointer as appropriate.

fn assert_panic(
ecx: &mut InterpCx<'mir, 'tcx, Self>,
msg: &AssertKind<Operand<'tcx>>,
unwind: Option<BasicBlock>
) -> Result<(), InterpErrorInfo<'tcx>>

Called to evaluate Assert MIR terminators that trigger a panic.

fn binary_ptr_op(
ecx: &InterpCx<'mir, 'tcx, Self>,
bin_op: BinOp,
left: &ImmTy<'tcx, Self::Provenance>,
right: &ImmTy<'tcx, Self::Provenance>
) -> Result<(Scalar<Self::Provenance>, bool, Ty<'tcx>), InterpErrorInfo<'tcx>>

Called for all binary operations where the LHS has pointer type.

Returns a (value, overflowed) pair if the operation succeeded

fn extern_static_base_pointer(
ecx: &InterpCx<'mir, 'tcx, Self>,
def_id: DefId
) -> Result<Pointer<Self::Provenance>, InterpErrorInfo<'tcx>>

Return the root pointer for the given extern static.

fn adjust_alloc_base_pointer(
ecx: &InterpCx<'mir, 'tcx, Self>,
ptr: Pointer<AllocId>
) -> Pointer<Self::Provenance>

Return a “base” pointer for the given allocation: the one that is used for direct accesses to this static/const/fn allocation, or the one returned from the heap allocator.

Not called on extern or thread-local statics (those use the methods above).

fn ptr_from_addr_cast(
ecx: &InterpCx<'mir, 'tcx, Self>,
addr: u64
) -> Result<Pointer<Option<Self::Provenance>>, InterpErrorInfo<'tcx>>

“Int-to-pointer cast”

fn expose_ptr(
ecx: &mut InterpCx<'mir, 'tcx, Self>,
ptr: Pointer<Self::Provenance>
) -> Result<(), InterpErrorInfo<'tcx>>

Marks a pointer as exposed, allowing it’s provenance to be recovered. “Pointer-to-int cast”

fn ptr_get_alloc(
ecx: &InterpCx<'mir, 'tcx, Self>,
ptr: Pointer<Self::Provenance>
) -> Option<(AllocId, Size, Self::ProvenanceExtra)>

Convert a pointer with provenance into an allocation-offset pair and extra provenance info.

The returned AllocId must be the same as ptr.provenance.get_alloc_id().

When this fails, that means the pointer does not point to a live allocation.

fn adjust_allocation(
ecx: &InterpCx<'mir, 'tcx, Self>,
id: AllocId,
alloc: Cow<'b, Allocation<AllocId, ()>>,
kind: Option<MemoryKind<Self::MemoryKind>>
) -> Result<Cow<'b, Allocation<Self::Provenance, Self::AllocExtra>>, InterpErrorInfo<'tcx>>

Called to adjust allocations to the Provenance and AllocExtra of this machine.

The way we construct allocations is to always first construct it without extra and then add the extra. This keeps uniform code paths for handling both allocations created by CTFE for globals, and allocations created by Miri during evaluation.

kind is the kind of the allocation being adjusted; it can be None when it’s a global and GLOBAL_KIND is None.

This should avoid copying if no work has to be done! If this returns an owned allocation (because a copy had to be done to adjust things), machine memory will cache the result. (This relies on AllocMap::get_or being able to add the owned allocation to the map even when the map is shared.)

This must only fail if alloc contains provenance.

fn init_frame_extra(
ecx: &mut InterpCx<'mir, 'tcx, Self>,
frame: Frame<'mir, 'tcx, Self::Provenance, ()>
) -> Result<Frame<'mir, 'tcx, Self::Provenance, Self::FrameExtra>, InterpErrorInfo<'tcx>>

Called immediately before a new stack frame gets pushed.

fn stack(
ecx: &'a InterpCx<'mir, 'tcx, Self>
) -> &'a [Frame<'mir, 'tcx, Self::Provenance, Self::FrameExtra>]

Borrow the current thread’s stack.

fn stack_mut(
ecx: &'a mut InterpCx<'mir, 'tcx, Self>
) -> &'a mut Vec<Frame<'mir, 'tcx, Self::Provenance, Self::FrameExtra>, Global>

Mutably borrow the current thread’s stack.

Provided Methods

fn enforce_abi(_ecx: &InterpCx<'mir, 'tcx, Self>) -> bool

Whether function calls should be ABI-checked.

fn load_mir(
ecx: &InterpCx<'mir, 'tcx, Self>,
instance: InstanceDef<'tcx>
) -> Result<&'tcx Body<'tcx>, InterpErrorInfo<'tcx>>

Entry point for obtaining the MIR of anything that should get evaluated. So not just functions and shims, but also const/static initializers, anonymous constants, …

fn abort(
_ecx: &mut InterpCx<'mir, 'tcx, Self>,
_msg: String
) -> Result<!, InterpErrorInfo<'tcx>>

Called to evaluate Abort MIR terminator.

fn access_local_mut(
ecx: &'a mut InterpCx<'mir, 'tcx, Self>,
frame: usize,
local: Local
) -> Result<&'a mut Operand<Self::Provenance>, InterpErrorInfo<'tcx>>where
'tcx: 'mir,

Called to write the specified local from the frame. Since writing a ZST is not actually accessing memory or locals, this is never invoked for ZST reads.

Due to borrow checker trouble, we indicate the frame as an index rather than an &mut Frame.

fn before_terminator(
_ecx: &mut InterpCx<'mir, 'tcx, Self>
) -> Result<(), InterpErrorInfo<'tcx>>

Called before a basic block terminator is executed. You can use this to detect endlessly running programs.

fn before_access_global(
_tcx: TyCtxt<'tcx>,
_machine: &Self,
_alloc_id: AllocId,
_allocation: ConstAllocation<'tcx, AllocId, ()>,
_static_def_id: Option<DefId>,
_is_write: bool
) -> Result<(), InterpErrorInfo<'tcx>>

Called before a global allocation is accessed. def_id is Some if this is the “lazy” allocation of a static.

fn thread_local_static_base_pointer(
_ecx: &mut InterpCx<'mir, 'tcx, Self>,
def_id: DefId
) -> Result<Pointer<Self::Provenance>, InterpErrorInfo<'tcx>>

Return the AllocId for the given thread-local static in the current thread.

fn eval_inline_asm(
_ecx: &mut InterpCx<'mir, 'tcx, Self>,
_template: &'tcx [InlineAsmTemplatePiece],
_operands: &[InlineAsmOperand<'tcx>],
_options: InlineAsmOptions
) -> Result<(), InterpErrorInfo<'tcx>>

fn before_memory_read(
_tcx: TyCtxt<'tcx>,
_machine: &Self,
_alloc_extra: &Self::AllocExtra,
_prov: (AllocId, Self::ProvenanceExtra),
_range: AllocRange
) -> Result<(), InterpErrorInfo<'tcx>>

Hook for performing extra checks on a memory read access.

Takes read-only access to the allocation so we can keep all the memory read operations take &self. Use a RefCell in AllocExtra if you need to mutate.

fn before_memory_write(
_tcx: TyCtxt<'tcx>,
_machine: &mut Self,
_alloc_extra: &mut Self::AllocExtra,
_prov: (AllocId, Self::ProvenanceExtra),
_range: AllocRange
) -> Result<(), InterpErrorInfo<'tcx>>

Hook for performing extra checks on a memory write access.

fn before_memory_deallocation(
_tcx: TyCtxt<'tcx>,
_machine: &mut Self,
_alloc_extra: &mut Self::AllocExtra,
_prov: (AllocId, Self::ProvenanceExtra),
_range: AllocRange
) -> Result<(), InterpErrorInfo<'tcx>>

Hook for performing extra operations on a memory deallocation.

fn retag(
_ecx: &mut InterpCx<'mir, 'tcx, Self>,
_kind: RetagKind,
_place: &PlaceTy<'tcx, Self::Provenance>
) -> Result<(), InterpErrorInfo<'tcx>>

Executes a retagging operation.

fn after_stack_push(
_ecx: &mut InterpCx<'mir, 'tcx, Self>
) -> Result<(), InterpErrorInfo<'tcx>>

Called immediately after a stack frame got pushed and its locals got initialized.

fn after_stack_pop(
_ecx: &mut InterpCx<'mir, 'tcx, Self>,
_frame: Frame<'mir, 'tcx, Self::Provenance, Self::FrameExtra>,
unwinding: bool
) -> Result<StackPopJump, InterpErrorInfo<'tcx>>

Called immediately after a stack frame got popped, but before jumping back to the caller. The locals have already been destroyed!