miri
Type Alias Scalar

pub type Scalar = Scalar<Provenance>;

Aliased Type

enum Scalar {
    Int(ScalarInt),
    Ptr(Pointer<Provenance>, u8),
}

Variants

Int(ScalarInt)

The raw bytes of a simple value.

Ptr(Pointer<Provenance>, u8)

A pointer.

We also store the size of the pointer, such that a Scalar always knows how big it is. The size is always the pointer size of the current target, but this is not information that we always have readily available.