miri

Module borrow_tracker

Modules

stacked_borrows

Implements “Stacked Borrows”. See https://github.com/rust-lang/unsafe-code-guidelines/blob/master/wip/stacked-borrows.md for further information.

tree_borrows

Structs

BorTag

Tracking pointer provenance

FrameState

Per-call-stack-frame data for borrow tracking

GlobalStateInner

Extra global state, available to the memory access hooks.

TreeBorrowsParams

Parameters that Tree Borrows can take.

Enums

AllocState

Extra per-allocation data for borrow tracking

BorrowTrackerMethod

Which borrow tracking method to use

ProtectorKind

The flavor of the protector.

RetagFields

Policy on whether to recurse into fields to retag

Traits

EvalContextExt

Type Aliases

GlobalState

We need interior mutable access to the global state.