miri::borrow_tracker::stacked_borrows
Type Alias AllocState

pub type AllocState = Stacks;

Aliased Type

struct AllocState {
    stacks: RangeMap<Stack>,
    history: AllocHistory,
    exposed_tags: HashSet<BorTag, FxBuildHasher>,
}

Fields
stacks: RangeMap<Stack>history: AllocHistory

Stores past operations on this allocation

exposed_tags: HashSet<BorTag, FxBuildHasher>

The set of tags that have been exposed inside this allocation.

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 232 bytes