Type Alias miri::borrow_tracker::stacked_borrows::AllocState

pub type AllocState = Stacks;

Aliased Type

struct AllocState {
    stacks: RangeMap<Stack>,
    history: AllocHistory,
    exposed_tags: HashSet<BorTag, BuildHasherDefault<FxHasher>>,
    modified_since_last_gc: bool,
}

Fields
stacks: RangeMap<Stack>history: AllocHistory

Stores past operations on this allocation

exposed_tags: HashSet<BorTag, BuildHasherDefault<FxHasher>>

The set of tags that have been exposed inside this allocation.

modified_since_last_gc: bool

Whether this memory has been modified since the last time the tag GC ran

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 240 bytes