miri::borrow_tracker
Struct GlobalStateInner

pub struct GlobalStateInner {
    borrow_tracker_method: BorrowTrackerMethod,
    next_ptr_tag: BorTag,
    root_ptr_tags: FxHashMap<AllocId, BorTag>,
    protected_tags: FxHashMap<BorTag, ProtectorKind>,
    tracked_pointer_tags: FxHashSet<BorTag>,
    retag_fields: RetagFields,
    unique_is_unique: bool,
}

Extra global state, available to the memory access hooks.

Fields
borrow_tracker_method: BorrowTrackerMethod

Borrow tracker method currently in use.

next_ptr_tag: BorTag

Next unused pointer ID (tag).

root_ptr_tags: FxHashMap<AllocId, BorTag>

Table storing the “root” tag for each allocation. The root tag is the one used for the initial pointer. We need this in a separate table to handle cyclic statics.

protected_tags: FxHashMap<BorTag, ProtectorKind>

All currently protected tags. We add tags to this when they are created with a protector in reborrow, and we remove tags from this when the call which is protecting them returns, in GlobalStateInner::end_call. See Stack::item_invalidated for more details.

tracked_pointer_tags: FxHashSet<BorTag>

The pointer ids to trace

retag_fields: RetagFields

Whether to recurse into datatypes when searching for pointers to retag.

unique_is_unique: bool

Whether core::ptr::Unique gets special (Box-like) handling.

Implementations

impl GlobalStateInner

Utilities for initialization and ID generation

pub fn new( borrow_tracker_method: BorrowTrackerMethod, tracked_pointer_tags: FxHashSet<BorTag>, retag_fields: RetagFields, unique_is_unique: bool, ) -> Self

fn new_ptr(&mut self) -> BorTag

Generates a new pointer tag. Remember to also check track_pointer_tags and log its creation!

pub fn new_frame(&mut self) -> FrameState

fn end_call(&mut self, frame: &FrameExtra<'_>)

pub fn root_ptr_tag(&mut self, id: AllocId, machine: &MiriMachine<'_>) -> BorTag

pub fn remove_unreachable_allocs(&mut self, allocs: &LiveAllocs<'_, '_>)

pub fn borrow_tracker_method(&self) -> BorrowTrackerMethod

impl GlobalStateInner

pub fn new_allocation( &mut self, id: AllocId, alloc_size: Size, kind: MemoryKind, machine: &MiriMachine<'_>, ) -> AllocState