Module miri::borrow_tracker::tree_borrows::diagnostics

Structs

DisplayFmt 🔒

All parameters to determine how the tree is formatted.

DisplayFmtAccess 🔒

How to show whether a location has been accessed

DisplayFmtPadding 🔒

Formatting of the tree structure.

DisplayFmtPermission 🔒

Formatting of the permissions on each range.

DisplayFmtWrapper 🔒

Pretty-printing details

DisplayIndent 🔒

Track the indentation of the tree.

DisplayRepr 🔒

Extracted information from the tree, in a form that is readily accessible for printing. I.e. resolve parent-child pointers into an actual tree, zip permissions with their tag, remove wrappers, stringify data.

Event

Complete data for an event:

History

List of all events that affected a tag. NOTE: not all of these events are relevant for a particular location, the events should be filtered before the generation of diagnostics. Available filtering methods include History::forget and History::extract_relevant.

HistoryData

History formatted for use by src/diagnostics.rs.

NodeDebugInfo

Some information that is irrelevant for the algorithm but very convenient to know about a tag for debugging and testing.

TbError 🔒

Failures that can occur during the execution of Tree Borrows procedures.

Enums

AccessCause

Cause of an access: either a real access or one inserted by Tree Borrows due to a reborrow or a deallocation.

TransitionError 🔒

Constants

DEFAULT_FORMATTER 🔒

Functions

char_repeat 🔒

Repeat a character a number of times.

Type Aliases

S 🔒