miri
Module clock

Structs

Clock

A monotone clock used for Instant simulation.

Instant

Enums

ClockKind 🔒

InstantKind 🔒

Constants

NANOSECONDS_PER_BASIC_BLOCK 🔒

When using a virtual clock, this defines how many nanoseconds we pretend are passing for each basic block. This number is pretty random, but it has been shown to approximately cause some sample programs to run within an order of magnitude of real time on desktop CPUs. (See tests/pass/shims/time-with-isolation*.rs.)