miri::concurrency::data_race
Struct MemoryCellClocks

struct MemoryCellClocks {
    write: (VectorIdx, VTimestamp),
    write_type: NaWriteType,
    read: VClock,
    atomic_ops: Option<Box<AtomicMemoryCellClocks>>,
}

Per-byte vector clock metadata for data-race detection.

Fields
write: (VectorIdx, VTimestamp)

The vector-clock timestamp and the thread that did the last non-atomic write. We donโt need a full VClock here, itโs always a single thread and nothing synchronizes, so the effective clock is all-0 except for the thread that did the write.

write_type: NaWriteType

The type of operation that the write index represents, either newly allocated memory, a non-atomic write or a deallocation of memory.

read: VClock

The vector-clock of all non-atomic reads that happened since the last non-atomic write (i.e., we join together the โsingletonโ clocks corresponding to each read). It is reset to zero on each write operation.

atomic_ops: Option<Box<AtomicMemoryCellClocks>>

Atomic access, acquire, release sequence tracking clocks. For non-atomic memory this value is set to None. For atomic memory, each byte carries this information.

Implementations

impl MemoryCellClocks

fn new(alloc: VTimestamp, alloc_index: VectorIdx) -> Self

Create a new set of clocks representing memory allocated at a given vector timestamp and index.

fn write_was_before(&self, other: &VClock) -> bool

fn write(&self) -> VClock

fn atomic(&self) -> Option<&AtomicMemoryCellClocks>

Load the internal atomic memory cells if they exist.

fn atomic_mut_unwrap(&mut self) -> &mut AtomicMemoryCellClocks

Load the internal atomic memory cells if they exist.

fn atomic_access( &mut self, thread_clocks: &ThreadClockSet, size: Size, write: bool, ) -> Result<&mut AtomicMemoryCellClocks, DataRace>

Load or create the internal atomic memory metadata if it does not exist. Also ensures we do not do mixed-size atomic accesses, and updates the recorded atomic access size.

fn load_acquire( &mut self, thread_clocks: &mut ThreadClockSet, index: VectorIdx, access_size: Size, ) -> Result<(), DataRace>

Update memory cell data-race tracking for atomic load acquire semantics, is a no-op if this memory was not used previously as atomic memory.

fn load_relaxed( &mut self, thread_clocks: &mut ThreadClockSet, index: VectorIdx, access_size: Size, ) -> Result<(), DataRace>

Update memory cell data-race tracking for atomic load relaxed semantics, is a no-op if this memory was not used previously as atomic memory.

fn store_release( &mut self, thread_clocks: &ThreadClockSet, index: VectorIdx, access_size: Size, ) -> Result<(), DataRace>

Update the memory cell data-race tracking for atomic store release semantics.

fn store_relaxed( &mut self, thread_clocks: &ThreadClockSet, index: VectorIdx, access_size: Size, ) -> Result<(), DataRace>

Update the memory cell data-race tracking for atomic store relaxed semantics.

fn rmw_release( &mut self, thread_clocks: &ThreadClockSet, index: VectorIdx, access_size: Size, ) -> Result<(), DataRace>

Update the memory cell data-race tracking for atomic store release semantics for RMW operations.

fn rmw_relaxed( &mut self, thread_clocks: &ThreadClockSet, index: VectorIdx, access_size: Size, ) -> Result<(), DataRace>

Update the memory cell data-race tracking for atomic store relaxed semantics for RMW operations.

fn atomic_read_detect( &mut self, thread_clocks: &ThreadClockSet, index: VectorIdx, access_size: Size, ) -> Result<(), DataRace>

Detect data-races with an atomic read, caused by a non-atomic write that does not happen-before the atomic-read.

fn atomic_write_detect( &mut self, thread_clocks: &ThreadClockSet, index: VectorIdx, access_size: Size, ) -> Result<(), DataRace>

Detect data-races with an atomic write, either with a non-atomic read or with a non-atomic write.

fn read_race_detect( &mut self, thread_clocks: &mut ThreadClockSet, index: VectorIdx, read_type: NaReadType, current_span: Span, ) -> Result<(), DataRace>

Detect races for non-atomic read operations at the current memory cell returns true if a data-race is detected.

fn write_race_detect( &mut self, thread_clocks: &mut ThreadClockSet, index: VectorIdx, write_type: NaWriteType, current_span: Span, ) -> Result<(), DataRace>

Detect races for non-atomic write operations at the current memory cell returns true if a data-race is detected.