miri::concurrency::data_race
Struct VClockAlloc

pub struct VClockAlloc {
    alloc_ranges: RefCell<RangeMap<MemoryCellClocks>>,
}

Vector clock metadata for a logical memory allocation.

Fields
alloc_ranges: RefCell<RangeMap<MemoryCellClocks>>

Assigning each byte a MemoryCellClocks.

Implementations

impl VClockAlloc

pub fn new_allocation( global: &GlobalState, thread_mgr: &ThreadManager<'_>, len: Size, kind: MemoryKind, current_span: Span, ) -> VClockAlloc

Create a new data-race detector for newly allocated memory.

fn find_gt_index(l: &VClock, r: &VClock) -> Option<VectorIdx>

fn report_data_race<'tcx>( global: &GlobalState, thread_mgr: &ThreadManager<'_>, mem_clocks: &MemoryCellClocks, access: AccessType, access_size: Size, ptr_dbg: Pointer<AllocId>, ty: Option<Ty<'_>>, ) -> InterpResult<'tcx>

Report a data-race found in the program. This finds the two racing threads and the type of data-race that occurred. This will also return info about the memory location the data-race occurred in. The ty parameter is used for diagnostics, letting the user know which type was involved in the access.

pub fn read<'tcx>( &self, alloc_id: AllocId, access_range: AllocRange, read_type: NaReadType, ty: Option<Ty<'_>>, machine: &MiriMachine<'_>, ) -> InterpResult<'tcx>

Detect data-races for an unsynchronized read operation. It will not perform data-race detection if race_detecting() is false, either due to no threads being created or if it is temporarily disabled during a racy read or write operation for which data-race detection is handled separately, for example atomic read operations. The ty parameter is used for diagnostics, letting the user know which type was read.

pub fn write<'tcx>( &mut self, alloc_id: AllocId, access_range: AllocRange, write_type: NaWriteType, ty: Option<Ty<'_>>, machine: &mut MiriMachine<'_>, ) -> InterpResult<'tcx>

Detect data-races for an unsynchronized write operation. It will not perform data-race detection if race_detecting() is false, either due to no threads being created or if it is temporarily disabled during a racy read or write operation. The ty parameter is used for diagnostics, letting the user know which type was written.