miri::concurrency::init_once
Trait EvalContextExt

pub trait EvalContextExt<'tcx>: MiriInterpCxExt<'tcx> {
    // Provided methods
    fn init_once_get_or_create_id(
        &mut self,
        lock: &MPlaceTy<'tcx>,
        offset: u64,
    ) -> InterpResult<'tcx, InitOnceId> { ... }
    fn init_once_status(&mut self, id: InitOnceId) -> InitOnceStatus { ... }
    fn init_once_enqueue_and_block(
        &mut self,
        id: InitOnceId,
        callback: impl UnblockCallback<'tcx> + 'tcx,
    ) { ... }
    fn init_once_begin(&mut self, id: InitOnceId) { ... }
    fn init_once_complete(&mut self, id: InitOnceId) -> InterpResult<'tcx> { ... }
    fn init_once_fail(&mut self, id: InitOnceId) -> InterpResult<'tcx> { ... }
    fn init_once_observe_completed(&mut self, id: InitOnceId) { ... }
}

Provided Methods

fn init_once_get_or_create_id( &mut self, lock: &MPlaceTy<'tcx>, offset: u64, ) -> InterpResult<'tcx, InitOnceId>

fn init_once_status(&mut self, id: InitOnceId) -> InitOnceStatus

fn init_once_enqueue_and_block( &mut self, id: InitOnceId, callback: impl UnblockCallback<'tcx> + 'tcx, )

Put the thread into the queue waiting for the initialization.

fn init_once_begin(&mut self, id: InitOnceId)

Begin initializing this InitOnce. Must only be called after checking that it is currently uninitialized.

fn init_once_complete(&mut self, id: InitOnceId) -> InterpResult<'tcx>

fn init_once_fail(&mut self, id: InitOnceId) -> InterpResult<'tcx>

fn init_once_observe_completed(&mut self, id: InitOnceId)

Synchronize with the previous completion of an InitOnce. Must only be called after checking that it is complete.

Object Safety

This trait is not object safe.