Module miri::eval

Main evaluator loop and setting up the initial stack frame.

Structs

MiriConfig

Configuration needed to spawn a Miri instance.

Enums

AlignmentCheck

BacktraceStyle

IsolatedOp

RejectOpWith

Functions

args_to_utf16_command_string 🔒

Turns an array of arguments into a Windows command line string.

create_ecx

Returns a freshly created InterpCx, along with an MPlaceTy representing the location where the return value of the start function will be written to. Public because this is also used by priroda.

eval_entry

Evaluates the entry function specified by entry_id. Returns Some(return_code) if program executed completed. Returns None if an evaluation error occurred.