Module miri::eval

Main evaluator loop and setting up the initial stack frame.

Structs

MiriConfig

Configuration needed to spawn a Miri instance.

Enums

AlignmentCheck

BacktraceStyle

IsolatedOp

MainThreadState 🔒

The state of the main thread. Implementation detail of on_main_stack_empty.

RejectOpWith

Constants

MAIN_THREAD_YIELDS_AT_SHUTDOWN 🔒

When the main thread would exit, we will yield to any other thread that is ready to execute. But we must only do that a finite number of times, or a background thread running loop {} will hang the program.

Functions

args_to_utf16_command_string 🔒

Turns an array of arguments into a Windows command line string.

create_ecx

Returns a freshly created InterpCx. Public because this is also used by priroda.

eval_entry

Evaluates the entry function specified by entry_id. Returns Some(return_code) if program executed completed. Returns None if an evaluation error occurred.