miri
Module helpers

Enums

AccessKind

Indicates which kind of access is being performed.

Traits

EvalContextExt

ToHost

Convert a softfloat type to its corresponding hostfloat type.

ToSoft

Convert a hostfloat type to its corresponding softfloat type.

Functions

bool_to_simd_element 🔒

check_arg_count

Check that the number of args is what we expect.

check_min_arg_count

Check that the number of args is at least the minumim what we expect.

get_local_crates

Retrieve the list of local crates that should have been passed by cargo-miri in MIRI_LOCAL_CRATES and turn them into CrateNums.

isolation_abort_error

iter_exported_symbols

Call f for each exported symbol.

path_ty_layout

Gets the layout of a type at a path.

resolve_path

Gets an instance for a path.

simd_element_to_bool 🔒

try_resolve_did 🔒

Gets an instance for a path.

try_resolve_path

Gets an instance for a path; fails gracefully if the path does not exist.

windows_check_buffer_size 🔒

Check whether an operation that writes to a target buffer was successful. Accordingly select return value. Local helper function to be used in Windows shims.