Module miri::helpers

Enums

AccessKind

Indicates which kind of access is being performed.

Constants

UNIX_IO_ERROR_TABLE 🔒

WINDOWS_IO_ERROR_TABLE 🔒

Traits

EvalContextExt

ToHost

Convert a softfloat type to its corresponding hostfloat type.

ToSoft

Convert a hostfloat type to its corresponding softfloat type.

Functions

bool_to_simd_element 🔒

check_arg_count

Check that the number of args is what we expect.

get_local_crates

Retrieve the list of local crates that should have been passed by cargo-miri in MIRI_LOCAL_CRATES and turn them into CrateNums.

isolation_abort_error

iter_exported_symbols

Call f for each exported symbol.

simd_element_to_bool 🔒

try_resolve_did 🔒

Gets an instance for a path.

windows_check_buffer_size 🔒

Check whether an operation that writes to a target buffer was successful. Accordingly select return value. Local helper function to be used in Windows shims.