Module miri::interpret

An interpreter for MIR used in CTFE and by miri

Macros

compile_time_machine

A lot of the flexibility above is just needed for Miri, but all “compile-time” machines (CTFE and ConstProp) use the same instance. Here, we share that code.

Structs

AllocDecodingSession

AllocDecodingState

AllocId

AllocRange

The information that makes up a memory access: offset and size.

AllocRef

A reference to some allocation that was already bounds-checked for the given region and had the on-access machine hooks run.

AllocRefMut

A reference to some allocation that was already bounds-checked for the given region and had the on-access machine hooks run.

Allocation

This type represents an Allocation in the Miri/CTFE core engine.

BadBytesAccess

Details of an access to uninitialized bytes / bad pointer bytes where it is not allowed.

ConstAlloc

Represents the result of const evaluation via the eval_to_allocation query.

ConstAllocation

Interned types generally have an Outer type and an Inner type, where Outer is a newtype around Interned<Inner>, and all the operations are done on Outer, because all occurrences are interned. E.g. Ty is an outer type and TyKind is its inner type.

Frame

A stack frame.

FrameInfo

What we store about a frame in an interpreter backtrace.

GlobalId

Uniquely identifies one of the following:

ImmTy

InitChunkIter

Yields InitChunks. See InitMask::range_as_init_chunks.

InterpCx

InterpErrorInfo

Packages the kind of error we got from the const code interpreter up with a Rust-level backtrace of where the error occurred. These should always be constructed by calling .into() on an InterpError. In rustc_mir::interpret, we have throw_err_* macros for this.

LitToConstInput

Input argument for tcx.lit_to_const.

LocalState

State of a local variable including a memoized layout

MPlaceTy

A MemPlace with its layout. Constructing it is only possible in this module.

MemPlace

Memory

OpTy

PlaceTy

Pointer

Represents a pointer in the Miri engine.

RefTracking

State for tracking recursive validation of references

ReportedErrorInfo

ScalarSizeMismatch

Information about a size mismatch.

ValidationErrorInfo

Enums

AllocError

We have our own error type that does not know about the AllocId; that information is added when converting to InterpError.

AllocKind

The return value of get_alloc_info indicates the “kind” of the allocation.

CheckInAllocMsg

Details of why a pointer had to be in-bounds.

ConstValue

Represents a constant value in Rust. Scalar and Slice are optimizations for array length computations, enum discriminants and the pattern matching logic.

CtfeValidationMode

Extra things to check for during validation of CTFE results.

ErrorHandled

ExpectedKind

FnArg

An argment passed to a function.

FnVal

The value of a function pointer.

GlobalAlloc

An allocation in the global (tcx-managed) memory can be either a function pointer, a static, or a “real” allocation with some data in it.

Immediate

An Immediate represents a single immediate self-contained Rust value.

InitChunk

A contiguous chunk of initialized or uninitialized memory.

InternKind

How a constant value should be interned.

InterpError

InvalidMetaKind

InvalidProgramInfo

Error information for when the program we executed turned out not to actually be a valid program. This cannot happen in stand-alone Miri, but it can happen during CTFE/ConstProp where we work on generic code or execution does not have all information available.

LitToConstError

Error type for tcx.lit_to_const.

LocalValue

Current value of a local variable

MemPlaceMeta

Information required for the sound usage of a MemPlace.

MemoryKind

Operand

An Operand is the result of computing a mir::Operand. It can be immediate, or still in memory. The latter is an optimization, to delay reading that chunk of memory and to avoid having to store arbitrary-sized data here.

Place

PointerKind

ResourceExhaustionInfo

Error information for when the program exhausted the resources granted to it by the interpreter.

Scalar

A Scalar represents an immediate, primitive value existing outside of a memory::Allocation. It is in many ways like a small chunk of an Allocation, up to 16 bytes in size. Like a range of bytes in an Allocation, a Scalar can either represent the raw bytes of a simple value or a pointer into another Allocation

StackPopCleanup

StackPopJump

Data returned by Machine::stack_pop, to provide further control over the popping of the stack frame

UndefinedBehaviorInfo

Error information for when the program caused Undefined Behavior.

UnsupportedOpInfo

Error information for when the program did something that might (or might not) be correct to do according to the Rust spec, but due to limitations in the interpreter, the operation could not be carried out. These limitations can differ between CTFE and the Miri engine, e.g., CTFE does not support dereferencing pointers at integral addresses.

ValidationErrorKind

Traits

AllocBytes

Functionality required for the bytes of an Allocation.

AllocMap

The functionality needed by memory to manage its allocations

Machine

Methods of this trait signifies a point where CTFE evaluation would fail and some use case dependent behaviour can instead be applied.

MachineStopType

A trait for machine-specific errors (or other “machine stop” conditions).

MayLeak

Whether this kind of memory is allowed to leak

PointerArithmetic

Projectable

A thing that we can project into, and that has a layout.

Provenance

This trait abstracts over the kind of provenance that is associated with a Pointer. It is mostly opaque; the Machine trait extends it with some more operations that also have access to some global state. The Debug rendering is used to display bare provenance, and for the default impl of fmt.

Readable

ValueVisitor

How to traverse a value and what to do when we are at the leaves.

Writeable

Functions

alloc_range

Free-starting constructor for less syntactic overhead.

get_slice_bytes

Gets the bytes of a constant slice value.

intern_const_alloc_recursive

Intern ret and everything it references.

read_target_uint

specialized_encode_alloc_id

struct_error

write_target_uint

Type Aliases

AllocResult

EvalToAllocationRawResult

EvalToConstValueResult

EvalToValTreeResult

InterpResult