miri::shims::env

Trait EvalContextExt

pub trait EvalContextExt<'tcx>: MiriInterpCxExt<'tcx> {
    // Provided methods
    fn get_env_var(
        &mut self,
        name: &OsStr,
    ) -> InterpResult<'tcx, Option<OsString>> { ... }
    fn get_pid(&self) -> u32 { ... }
    fn get_current_tid(&self) -> u32 { ... }
    fn get_tid(&self, thread: ThreadId) -> u32 { ... }
}

Provided Methods

fn get_env_var(&mut self, name: &OsStr) -> InterpResult<'tcx, Option<OsString>>

Try to get an environment variable from the interpreted program’s environment. This is useful for implementing shims which are documented to read from the environment.

fn get_pid(&self) -> u32

Get the process identifier.

fn get_current_tid(&self) -> u32

Get an “OS” thread ID for the current thread.

fn get_tid(&self, thread: ThreadId) -> u32

Get an “OS” thread ID for any thread.