Trait miri::shims::native_lib::EvalContextExtPriv

trait EvalContextExtPriv<'tcx>: MiriInterpCxExt<'tcx> {
    // Provided methods
    fn call_native_with_args<'a>(
        &mut self,
        link_name: Symbol,
        dest: &MPlaceTy<'tcx>,
        ptr: CodePtr,
        libffi_args: Vec<Arg<'a>>,
    ) -> InterpResult<'tcx, ImmTy<'tcx>> { ... }
    fn get_func_ptr_explicitly_from_lib(
        &mut self,
        link_name: Symbol,
    ) -> Option<CodePtr> { ... }
}

Provided Methods

fn call_native_with_args<'a>( &mut self, link_name: Symbol, dest: &MPlaceTy<'tcx>, ptr: CodePtr, libffi_args: Vec<Arg<'a>>, ) -> InterpResult<'tcx, ImmTy<'tcx>>

Call native host function and return the output as an immediate.

fn get_func_ptr_explicitly_from_lib( &mut self, link_name: Symbol, ) -> Option<CodePtr>

Get the pointer to the function of the specified name in the shared object file, if it exists. The function must be in the shared object file specified: we do not return pointers to functions in dependencies of the library.