miri::shims::panic

Trait EvalContextExt

pub trait EvalContextExt<'tcx>: MiriInterpCxExt<'tcx> {
    // Provided methods
    fn start_panic(
        &mut self,
        msg: &str,
        unwind: UnwindAction,
    ) -> InterpResult<'tcx> { ... }
    fn start_panic_nounwind(&mut self, msg: &str) -> InterpResult<'tcx> { ... }
    fn assert_panic(
        &mut self,
        msg: &AssertMessage<'tcx>,
        unwind: UnwindAction,
    ) -> InterpResult<'tcx> { ... }
}

Provided Methods

fn start_panic(&mut self, msg: &str, unwind: UnwindAction) -> InterpResult<'tcx>

Start a panic in the interpreter with the given message as payload.

fn start_panic_nounwind(&mut self, msg: &str) -> InterpResult<'tcx>

Start a non-unwinding panic in the interpreter with the given message as payload.

fn assert_panic( &mut self, msg: &AssertMessage<'tcx>, unwind: UnwindAction, ) -> InterpResult<'tcx>