miri::shims::unix::fd
Trait EvalContextExt

pub trait EvalContextExt<'tcx>: MiriInterpCxExt<'tcx> {
    // Provided methods
    fn dup(&mut self, old_fd_num: i32) -> InterpResult<'tcx, Scalar> { ... }
    fn dup2(
        &mut self,
        old_fd_num: i32,
        new_fd_num: i32,
    ) -> InterpResult<'tcx, Scalar> { ... }
    fn flock(&mut self, fd_num: i32, op: i32) -> InterpResult<'tcx, Scalar> { ... }
    fn fcntl(&mut self, args: &[OpTy<'tcx>]) -> InterpResult<'tcx, Scalar> { ... }
    fn close(&mut self, fd_op: &OpTy<'tcx>) -> InterpResult<'tcx, Scalar> { ... }
    fn read(
        &mut self,
        fd_num: i32,
        buf: Pointer,
        count: u64,
        offset: Option<i128>,
        dest: &MPlaceTy<'tcx>,
    ) -> InterpResult<'tcx> { ... }
    fn write(
        &mut self,
        fd_num: i32,
        buf: Pointer,
        count: u64,
        offset: Option<i128>,
        dest: &MPlaceTy<'tcx>,
    ) -> InterpResult<'tcx> { ... }
}

Provided Methods

fn dup(&mut self, old_fd_num: i32) -> InterpResult<'tcx, Scalar>

fn dup2( &mut self, old_fd_num: i32, new_fd_num: i32, ) -> InterpResult<'tcx, Scalar>

fn flock(&mut self, fd_num: i32, op: i32) -> InterpResult<'tcx, Scalar>

fn fcntl(&mut self, args: &[OpTy<'tcx>]) -> InterpResult<'tcx, Scalar>

fn close(&mut self, fd_op: &OpTy<'tcx>) -> InterpResult<'tcx, Scalar>

fn read( &mut self, fd_num: i32, buf: Pointer, count: u64, offset: Option<i128>, dest: &MPlaceTy<'tcx>, ) -> InterpResult<'tcx>

Read data from fd into buffer specified by buf and count.

If offset is None, reads data from current cursor position associated with fd and updates cursor position on completion. Otherwise, reads from the specified offset and keeps the cursor unchanged.

fn write( &mut self, fd_num: i32, buf: Pointer, count: u64, offset: Option<i128>, dest: &MPlaceTy<'tcx>, ) -> InterpResult<'tcx>