miri::shims::unix::linux_like::epoll
Struct Epoll

struct Epoll {
    interest_list: RefCell<BTreeMap<(FdId, i32), Rc<RefCell<EpollEventInterest>>>>,
    ready_list: Rc<ReadyList>,
    thread_id: RefCell<Vec<ThreadId>>,
}

An Epoll file descriptor connects file handles and epoll events

Fields
interest_list: RefCell<BTreeMap<(FdId, i32), Rc<RefCell<EpollEventInterest>>>>

A map of EpollEventInterests registered under this epoll instance. Each entry is differentiated using FdId and file descriptor value.

ready_list: Rc<ReadyList>

A map of EpollEventInstance that will be returned when epoll_wait is called. Similar to interest_list, the entry is also differentiated using FdId and file descriptor value.

thread_id: RefCell<Vec<ThreadId>>

A list of thread ids blocked on this epoll instance.

Implementations

impl Epoll

fn get_ready_list(&self) -> Rc<ReadyList>