miri::shims::unix::linux_like
Module eventfd

Linux eventfd implementation.

Structs

Event 🔒

A kind of file descriptor created by eventfd. The Event type isn’t currently written to by eventfd. The interface is meant to keep track of objects associated with a file descriptor. For more information see the man page below:

Constants

MAX_COUNTER 🔒

Maximum value that the eventfd counter can hold.

Traits

EvalContextExt

Functions

eventfd_read 🔒

Block thread if the current counter is 0, else just return the current counter value to the caller and set the counter to 0.

eventfd_write 🔒

Block thread if the value addition will exceed u64::MAX -1, else just add the user-supplied value to current counter.