Module miri::shims::unix::sync

Constants

PTHREAD_MUTEX_NORMAL_FLAG ğŸ”’

A flag that allows to distinguish PTHREAD_MUTEX_NORMAL from PTHREAD_MUTEX_DEFAULT. Since in glibc they have the same numeric values, but different behaviour, we need a way to distinguish them. We do this by setting this bit flag to the PTHREAD_MUTEX_NORMAL mutexes. See the comment in pthread_mutexattr_settype function.

Traits

EvalContextExt

Functions

cond_get_clock_id ğŸ”’

cond_get_id ğŸ”’

cond_get_or_create_id ğŸ”’

cond_set_clock_id ğŸ”’

cond_set_id ğŸ”’

condattr_get_clock_id ğŸ”’

condattr_set_clock_id ğŸ”’

is_mutex_kind_default ğŸ”’

is_mutex_kind_normal ğŸ”’

layout_of_maybe_uninit ğŸ”’

mutex_get_id ğŸ”’

mutex_get_kind ğŸ”’

mutex_get_or_create_id ğŸ”’

mutex_set_id ğŸ”’

mutex_set_kind ğŸ”’

mutexattr_get_kind ğŸ”’

mutexattr_set_kind ğŸ”’

post_cond_signal ğŸ”’

After a thread waiting on a condvar was signalled: Reacquire the conditional variable and remove the timeout callback if any was registered.

reacquire_cond_mutex ğŸ”’

Try to reacquire the mutex associated with the condition variable after we were signaled.

release_cond_mutex_and_block ğŸ”’

Release the mutex associated with the condition variable because we are entering the waiting state.

rwlock_get_id ğŸ”’

rwlock_get_or_create_id ğŸ”’