Module miri::shims::unix::sync

Constants

PTHREAD_MUTEX_NORMAL_FLAG ğŸ”’

A flag that allows to distinguish PTHREAD_MUTEX_NORMAL from PTHREAD_MUTEX_DEFAULT. Since in glibc they have the same numeric values, but different behaviour, we need a way to distinguish them. We do this by setting this bit flag to the PTHREAD_MUTEX_NORMAL mutexes. See the comment in pthread_mutexattr_settype function.

Traits

EvalContextExt

Functions

cond_clock_offset ğŸ”’

cond_get_clock_id ğŸ”’

cond_get_id ğŸ”’

cond_id_offset ğŸ”’

cond_reset_id ğŸ”’

cond_set_clock_id ğŸ”’

condattr_clock_offset ğŸ”’

condattr_get_clock_id ğŸ”’

condattr_set_clock_id ğŸ”’

is_cond_clock_realtime ğŸ”’

Determines whether this clock represents the real-time clock, CLOCK_REALTIME.

is_mutex_kind_default ğŸ”’

is_mutex_kind_normal ğŸ”’

mutex_get_id ğŸ”’

mutex_get_kind ğŸ”’

mutex_id_offset ğŸ”’

mutex_kind_offset ğŸ”’

mutex_reset_id ğŸ”’

mutex_set_kind ğŸ”’

mutexattr_get_kind ğŸ”’

mutexattr_kind_offset ğŸ”’

mutexattr_set_kind ğŸ”’

rwlock_get_id ğŸ”’

rwlock_id_offset ğŸ”’