miri::shims::x86
Function split_simd_to_128bit_chunks

fn split_simd_to_128bit_chunks<'tcx, P: Projectable<'tcx, Provenance>>(
    ecx: &mut MiriInterpCx<'tcx>,
    op: &P,
) -> InterpResult<'tcx, (u64, u64, P)>

Splits op (which must be a SIMD vector) into 128-bit chunks.

Returns a tuple where:

The first element is the number of 128-bit chunks (let’s call it N).

The second element is the number of elements per chunk (let’s call it M).

The third element is the op vector split into chunks, i.e, it’s type is [[T; M]; N] where T is the element type of op.