miri::shims::x86
Module sse42

Constants

USE_SIGNED 🔒

A bitmask constant for scrutinizing the immediate byte provided to the string comparison intrinsics. It distinuishes between signed integers and unsigned integers. See compare_strings for more details about the immediate byte.

USE_WORDS 🔒

A bitmask constant for scrutinizing the immediate byte provided to the string comparison intrinsics. It distinuishes between 16-bit integers and 8-bit integers. See compare_strings for more details about the immediate byte.

Traits

EvalContextExt 🔒

Functions

compare_strings 🔒

The main worker for the string comparison intrinsics, where the given strings are analyzed according to the given immediate byte.

deconstruct_args 🔒

Obtain the arguments of the intrinsic based on its name. The result is a tuple with the following values:

default_len 🔒

implicit_len 🔒

Calculate the c-style string length for a given string str. The string is either a length 16 array of bytes a length 8 array of two-byte words.