miri::shims::x86::sse42
Function deconstruct_args

fn deconstruct_args<'tcx>(
    unprefixed_name: &str,
    ecx: &mut MiriInterpCx<'tcx>,
    link_name: Symbol,
    abi: ExternAbi,
    args: &[OpTy<'tcx>],
) -> InterpResult<'tcx, (OpTy<'tcx>, OpTy<'tcx>, Option<(u64, u64)>, u8)>

Obtain the arguments of the intrinsic based on its name. The result is a tuple with the following values:

The first string argument.

The second string argument.

The string length values, if the intrinsic requires them.

The immediate instruction byte.

The string arguments will be transmuted into arrays of bytes or two-byte words, depending on the value of the immediate byte. Originally, they are __m128i values corresponding to the x86 128-bit integer SIMD type.