Struct miri::MiriMachine

pub struct MiriMachine<'mir, 'tcx> {
    pub tcx: TyCtxt<'tcx>,
    pub borrow_tracker: Option<RefCell<GlobalStateInner>>,
    pub data_race: Option<GlobalState>,
    pub alloc_addresses: RefCell<GlobalStateInner>,
    pub native_lib: Option<!>,
    /* private fields */
}

The machine itself.

If you add anything here that stores machine values, remember to update visit_all_machine_values!

Fields
tcx: TyCtxt<'tcx>borrow_tracker: Option<RefCell<GlobalStateInner>>

Global data for borrow tracking.

data_race: Option<GlobalState>

Data race detector global data.

alloc_addresses: RefCell<GlobalStateInner>

Ptr-int-cast module global data.

native_lib: Option<!>
Implementations

impl<'mir, 'tcx> MiriMachine<'mir, 'tcx>

pub fn free_alloc_id( &mut self, dead_id: AllocId, size: Size, align: Align, kind: MemoryKind )

impl<'mir, 'tcx> MiriMachine<'mir, 'tcx>

pub fn emit_diagnostic(&self, e: NonHaltingDiagnostic)

impl<'mir, 'tcx> MiriMachine<'mir, 'tcx>

pub fn current_span(&self) -> Span

Get the current span in the topmost function which is workspace-local and not #[track_caller]. This function is backed by a cache, and can be assumed to be very fast. It will work even when the stack is empty.

pub fn caller_span(&self) -> Span

Returns the span of the caller of the current operation, again walking down the stack to find the closest frame in a local crate, if the caller of the current operation is not in a local crate. This is useful when we are processing something which occurs on function-entry and we want to point at the call to the function, not the function definition generally.

pub fn is_user_relevant(&self, frame: &Frame<'mir, 'tcx, Provenance>) -> bool

This is the source of truth for the is_user_relevant flag in our FrameExtra.

impl<'mir, 'tcx> MiriMachine<'mir, 'tcx>

pub fn init_extern_statics( this: &mut MiriInterpCx<'mir, 'tcx> ) -> InterpResult<'tcx>

Sets up the “extern statics” for this machine.