miri

Struct MiriMachine

pub struct MiriMachine<'tcx> {

    pub tcx: TyCtxt<'tcx>,
    pub borrow_tracker: Option<RefCell<GlobalStateInner>>,
    pub data_race: GlobalDataRaceHandler,
    pub alloc_addresses: RefCell<GlobalStateInner>,
    pub(crate) env_vars: EnvVars<'tcx>,
    pub(crate) main_fn_ret_place: Option<MPlaceTy<'tcx>>,
    pub(crate) argc: Option<Pointer>,
    pub(crate) argv: Option<Pointer>,
    pub(crate) cmd_line: Option<Pointer>,
    pub(crate) tls: TlsData<'tcx>,
    pub(crate) isolated_op: IsolatedOp,
    pub(crate) validation: ValidationMode,
    pub(crate) fds: FdTable,
    pub(crate) dirs: DirTable,
    pub(crate) epoll_interests: EpollInterestTable,
    pub(crate) monotonic_clock: MonotonicClock,
    pub(crate) threads: ThreadManager<'tcx>,
    pub(crate) thread_cpu_affinity: FxHashMap<ThreadId, CpuAffinityMask>,
    pub(crate) layouts: PrimitiveLayouts<'tcx>,
    pub(crate) static_roots: Vec<AllocId>,
    profiler: Option<Profiler>,
    string_cache: FxHashMap<String, StringId>,
    pub(crate) exported_symbols_cache: FxHashMap<Symbol, Option<Instance<'tcx>>>,
    pub(crate) backtrace_style: BacktraceStyle,
    pub(crate) local_crates: Vec<CrateNum>,
    extern_statics: FxHashMap<Symbol, StrictPointer>,
    pub(crate) rng: RefCell<StdRng>,
    pub(crate) allocator: Option<Rc<RefCell<IsolatedAlloc>>>,
    tracked_alloc_ids: FxHashSet<AllocId>,
    track_alloc_accesses: bool,
    pub(crate) check_alignment: AlignmentCheck,
    pub(crate) cmpxchg_weak_failure_rate: f64,
    pub(crate) preemption_rate: f64,
    pub(crate) report_progress: Option<u32>,
    pub(crate) basic_block_count: u64,
    pub native_lib: Vec<(Library, PathBuf)>,
    pub(crate) gc_interval: u32,
    pub(crate) since_gc: u32,
    pub(crate) num_cpus: u32,
    pub(crate) page_size: u64,
    pub(crate) stack_addr: u64,
    pub(crate) stack_size: u64,
    pub(crate) collect_leak_backtraces: bool,
    pub(crate) allocation_spans: RefCell<FxHashMap<AllocId, (Span, Option<Span>)>>,
    pub(crate) symbolic_alignment: RefCell<FxHashMap<AllocId, (Size, Align)>>,
    union_data_ranges: FxHashMap<Ty<'tcx>, RangeSet>,
    pub(crate) pthread_mutex_sanity: Cell<bool>,
    pub(crate) pthread_rwlock_sanity: Cell<bool>,
    pub(crate) pthread_condvar_sanity: Cell<bool>,
    pub(crate) sb_extern_type_warned: Cell<bool>,
    pub(crate) native_call_mem_warned: Cell<bool>,
    pub(crate) reject_in_isolation_warned: RefCell<FxHashSet<String>>,
    pub(crate) int2ptr_warned: RefCell<FxHashSet<Span>>,
    pub(crate) mangle_internal_symbol_cache: FxHashMap<&'static str, String>,
    pub force_intrinsic_fallback: bool,
    pub float_nondet: bool,
    pub float_rounding_error: bool,

}

The machine itself.

If you add anything here that stores machine values, remember to update visit_all_machine_values!

Fields
tcx: TyCtxt<'tcx>borrow_tracker: Option<RefCell<GlobalStateInner>>

Global data for borrow tracking.

data_race: GlobalDataRaceHandler

Depending on settings, this will be None, global data for a data race detector, or the context required for running in GenMC mode.

Invariant: The enum variant must match the enum variant of AllocDataRaceHandler in the data_race field of all AllocExtra.

alloc_addresses: RefCell<GlobalStateInner>

Ptr-int-cast module global data.

env_vars: EnvVars<'tcx>

Environment variables.

main_fn_ret_place: Option<MPlaceTy<'tcx>>

Return place of the main function.

argc: Option<Pointer>

Program arguments (Option because we can only initialize them after creating the ecx). These are pointers to argc/argv because macOS. We also need the full command line as one string because of Windows.

argv: Option<Pointer>cmd_line: Option<Pointer>tls: TlsData<'tcx>

TLS state.

isolated_op: IsolatedOp

What should Miri do when an op requires communicating with the host, such as accessing host env vars, random number generation, and file system access.

validation: ValidationMode

Whether to enforce the validity invariant.

fds: FdTable

The table of file descriptors.

dirs: DirTable

The table of directory descriptors.

epoll_interests: EpollInterestTable

The list of all EpollEventInterest.

monotonic_clock: MonotonicClock

This machine’s monotone clock.

threads: ThreadManager<'tcx>

The set of threads.

thread_cpu_affinity: FxHashMap<ThreadId, CpuAffinityMask>

Stores which thread is eligible to run on which CPUs. This has no effect at all, it is just tracked to produce the correct result in sched_getaffinity

layouts: PrimitiveLayouts<'tcx>

Precomputed TyLayouts for primitive data types that are commonly used inside Miri.

static_roots: Vec<AllocId>

Allocations that are considered roots of static memory (that may leak).

profiler: Option<Profiler>

The measureme profiler used to record timing information about the emulated program.

string_cache: FxHashMap<String, StringId>

Used with profiler to cache the StringIds for event names used with measureme.

exported_symbols_cache: FxHashMap<Symbol, Option<Instance<'tcx>>>

Cache of Instance exported under the given Symbol name. None means no Instance exported under the given name is found.

backtrace_style: BacktraceStyle

Equivalent setting as RUST_BACKTRACE on encountering an error.

local_crates: Vec<CrateNum>

Crates which are considered local for the purposes of error reporting.

extern_statics: FxHashMap<Symbol, StrictPointer>

Mapping extern static names to their pointer.

rng: RefCell<StdRng>

The random number generator used for resolving non-determinism. Needs to be queried by ptr_to_int, hence needs interior mutability.

allocator: Option<Rc<RefCell<IsolatedAlloc>>>

The allocator used for the machine’s AllocBytes in native-libs mode.

tracked_alloc_ids: FxHashSet<AllocId>

The allocation IDs to report when they are being allocated (helps for debugging memory leaks and use after free bugs).

track_alloc_accesses: bool

For the tracked alloc ids, also report read/write accesses.

check_alignment: AlignmentCheck

Controls whether alignment of memory accesses is being checked.

cmpxchg_weak_failure_rate: f64

Failure rate of compare_exchange_weak, between 0.0 and 1.0

preemption_rate: f64

The probability of the active thread being preempted at the end of each basic block.

report_progress: Option<u32>

If Some, we will report the current stack every N basic blocks.

basic_block_count: u64native_lib: Vec<(Library, PathBuf)>

Handle of the optional shared object file for native functions.

gc_interval: u32

Run a garbage collector for BorTags every N basic blocks.

since_gc: u32

The number of blocks that passed since the last BorTag GC pass.

num_cpus: u32

The number of CPUs to be reported by miri.

page_size: u64

Determines Miri’s page size and associated values

stack_addr: u64stack_size: u64collect_leak_backtraces: bool

Whether to collect a backtrace when each allocation is created, just in case it leaks.

allocation_spans: RefCell<FxHashMap<AllocId, (Span, Option<Span>)>>

The spans we will use to report where an allocation was created and deallocated in diagnostics.

symbolic_alignment: RefCell<FxHashMap<AllocId, (Size, Align)>>

For each allocation, an offset inside that allocation that was deemed aligned even for symbolic alignment checks. This cannot be stored in AllocExtra since it needs to be tracked for vtables and function allocations as well as regular allocations.

Invariant: the promised alignment will never be less than the native alignment of the allocation.

union_data_ranges: FxHashMap<Ty<'tcx>, RangeSet>

A cache of “data range” computations for unions (i.e., the offsets of non-padding bytes).

pthread_mutex_sanity: Cell<bool>

Caches the sanity-checks for various pthread primitives.

pthread_rwlock_sanity: Cell<bool>pthread_condvar_sanity: Cell<bool>sb_extern_type_warned: Cell<bool>

Remembers whether we already warned about an extern type with Stacked Borrows.

native_call_mem_warned: Cell<bool>

Remember whether we already warned about sharing memory with a native call.

reject_in_isolation_warned: RefCell<FxHashSet<String>>

Remembers which shims have already shown the warning about erroring in isolation.

int2ptr_warned: RefCell<FxHashSet<Span>>

Remembers which int2ptr casts we have already warned about.

mangle_internal_symbol_cache: FxHashMap<&'static str, String>

Cache for mangle_internal_symbol.

force_intrinsic_fallback: bool

Always prefer the intrinsic fallback body over the native Miri implementation.

float_nondet: bool

Whether floating-point operations can behave non-deterministically.

float_rounding_error: bool

Whether floating-point operations can have a non-deterministic rounding error.

Implementations

impl<'tcx> MiriMachine<'tcx>

pub fn free_alloc_id( &mut self, dead_id: AllocId, size: Size, align: Align, kind: MemoryKind, )

impl<'tcx> MiriMachine<'tcx>

pub fn emit_diagnostic(&self, e: NonHaltingDiagnostic)

impl<'tcx> MiriMachine<'tcx>

pub fn current_span(&self) -> Span

Get the current span in the topmost function which is workspace-local and not #[track_caller]. This function is backed by a cache, and can be assumed to be very fast. It will work even when the stack is empty.

pub fn caller_span(&self) -> Span

Returns the span of the caller of the current operation, again walking down the stack to find the closest frame in a local crate, if the caller of the current operation is not in a local crate. This is useful when we are processing something which occurs on function-entry and we want to point at the call to the function, not the function definition generally.

fn stack(&self) -> &[Frame<'tcx, Provenance, FrameExtra<'tcx>>]

fn top_user_relevant_frame(&self) -> Option<usize>

pub fn is_user_relevant(&self, frame: &Frame<'tcx, Provenance>) -> bool

This is the source of truth for the is_user_relevant flag in our FrameExtra.

impl<'tcx> MiriMachine<'tcx>

pub(crate) fn new( config: &MiriConfig, layout_cx: LayoutCx<'tcx>, genmc_ctx: Option<Rc<GenmcCtx>>, ) -> Self

Create a new MiriMachine.

Invariant: genmc_ctx.is_some() == config.genmc_config.is_some()

pub(crate) fn late_init( ecx: &mut MiriInterpCx<'tcx>, config: &MiriConfig, on_main_stack_empty: StackEmptyCallback<'tcx>, ) -> InterpResult<'tcx>

pub(crate) fn add_extern_static( ecx: &mut MiriInterpCx<'tcx>, name: &str, ptr: Pointer, )

pub(crate) fn communicate(&self) -> bool

pub(crate) fn is_local(&self, frame: &FrameInfo<'_>) -> bool

Check whether the stack frame that this FrameInfo refers to is part of a local crate.

pub(crate) fn handle_abnormal_termination(&mut self)

Called when the interpreter is going to shut down abnormally, such as due to a Ctrl-C.

pub(crate) fn page_align(&self) -> Align

pub(crate) fn allocated_span(&self, alloc_id: AllocId) -> Option<SpanData>

pub(crate) fn deallocated_span(&self, alloc_id: AllocId) -> Option<SpanData>

fn init_allocation( ecx: &MiriInterpCx<'tcx>, id: AllocId, kind: MemoryKind, size: Size, align: Align, ) -> InterpResult<'tcx, AllocExtra<'tcx>>

impl<'tcx> MiriMachine<'tcx>

fn alloc_extern_static( ecx: &mut MiriInterpCx<'tcx>, name: &str, val: ImmTy<'tcx>, ) -> InterpResult<'tcx>

fn null_ptr_extern_statics( ecx: &mut MiriInterpCx<'tcx>, names: &[&str], ) -> InterpResult<'tcx>

Zero-initialized pointer-sized extern statics are pretty common. Most of them are for weak symbols, which we all set to null (indicating that the symbol is not supported, and triggering fallback code which ends up calling some other shim that we do support).

fn weak_symbol_extern_statics( ecx: &mut MiriInterpCx<'tcx>, names: &[&str], ) -> InterpResult<'tcx>

Extern statics that are initialized with function pointers to the symbols of the same name.

pub fn init_extern_statics(ecx: &mut MiriInterpCx<'tcx>) -> InterpResult<'tcx>

Sets up the “extern statics” for this machine.