Struct miri::ThreadManager

pub struct ThreadManager<'mir, 'tcx> { /* private fields */ }

A set of threads.

Implementations

impl<'mir, 'tcx: 'mir> ThreadManager<'mir, 'tcx>

pub fn active_thread_stack( &self ) -> &[Frame<'mir, 'tcx, Provenance, FrameExtra<'tcx>>]

Borrow the stack of the active thread.

pub fn all_stacks( &self ) -> impl Iterator<Item = &[Frame<'mir, 'tcx, Provenance, FrameExtra<'tcx>>]>

pub fn get_active_thread_id(&self) -> ThreadId

Get the id of the currently active thread.

pub fn get_total_thread_count(&self) -> usize

Get the total number of threads that were ever spawn by this program.

pub fn get_live_thread_count(&self) -> usize

Get the total of threads that are currently live, i.e., not yet terminated. (They might be blocked.)

pub fn active_thread_mut(&mut self) -> &mut Thread<'mir, 'tcx>

Get a mutable borrow of the currently active thread.

pub fn active_thread_ref(&self) -> &Thread<'mir, 'tcx>

Get a shared borrow of the currently active thread.

pub fn set_thread_name(&mut self, thread: ThreadId, new_thread_name: Vec<u8>)

Set the name of the given thread.

pub fn get_thread_name(&self, thread: ThreadId) -> &[u8]

Get the name of the given thread.